@@ -3,7 +3,7 @@
 
 #include <toml.hpp>
 
-#include "../handler/settings.h"
+#include "handler/settings.h"
 #include "crontask.h"
 #include "proxygroup.h"
 #include "regmatch.h"
@@ -17,9 +17,9 @@ namespace toml
         static ProxyGroupConfig from_toml(const value& v)
         {
             ProxyGroupConfig conf;
-            conf.Name = toml::find<String>(v, "name");
-            String type = toml::find<String>(v, "type");
-            String strategy = toml::find_or<String>(v, "strategy", "");
+            conf.Name = find<String>(v, "name");
+            String type = find<String>(v, "type");
+            String strategy = find_or<String>(v, "strategy", "");
             switch(hash_(type))
             {
             case "select"_hash:
@@ -27,18 +27,18 @@ namespace toml
                 break;
             case "url-test"_hash:
                 conf.Type = ProxyGroupType::URLTest;
-                conf.Url = toml::find<String>(v, "url");
-                conf.Interval = toml::find<Integer>(v, "interval");
-                conf.Tolerance = toml::find_or<Integer>(v, "tolerance", 0);
+                conf.Url = find<String>(v, "url");
+                conf.Interval = find<Integer>(v, "interval");
+                conf.Tolerance = find_or<Integer>(v, "tolerance", 0);
                 if(v.contains("lazy"))
-                    conf.Lazy = toml::find_or<bool>(v, "lazy", false);
+                    conf.Lazy = find_or<bool>(v, "lazy", false);
                 if(v.contains("evaluate-before-use"))
-                    conf.EvaluateBeforeUse = toml::find_or(v, "evaluate-before-use", conf.EvaluateBeforeUse.get());
+                    conf.EvaluateBeforeUse = find_or(v, "evaluate-before-use", conf.EvaluateBeforeUse.get());
                 break;
             case "load-balance"_hash:
                 conf.Type = ProxyGroupType::LoadBalance;
-                conf.Url = toml::find<String>(v, "url");
-                conf.Interval = toml::find<Integer>(v, "interval");
+                conf.Url = find<String>(v, "url");
+                conf.Interval = find<Integer>(v, "interval");
                 switch(hash_(strategy))
                 {
                 case "consistent-hashing"_hash:
@@ -49,14 +49,14 @@ namespace toml
                     break;
                 }
                 if(v.contains("persistent"))
-                    conf.Persistent = toml::find_or(v, "persistent", conf.Persistent.get());
+                    conf.Persistent = find_or(v, "persistent", conf.Persistent.get());
                 break;
             case "fallback"_hash:
                 conf.Type = ProxyGroupType::Fallback;
-                conf.Url = toml::find<String>(v, "url");
-                conf.Interval = toml::find<Integer>(v, "interval");
+                conf.Url = find<String>(v, "url");
+                conf.Interval = find<Integer>(v, "interval");
                 if(v.contains("evaluate-before-use"))
-                    conf.EvaluateBeforeUse = toml::find_or(v, "evaluate-before-use", conf.EvaluateBeforeUse.get());
+                    conf.EvaluateBeforeUse = find_or(v, "evaluate-before-use", conf.EvaluateBeforeUse.get());
                 break;
             case "relay"_hash:
                 conf.Type = ProxyGroupType::Relay;
@@ -64,16 +64,26 @@ namespace toml
             case "ssid"_hash:
                 conf.Type = ProxyGroupType::SSID;
                 break;
+            case "smart"_hash:
+                conf.Type = ProxyGroupType::Smart;
+                conf.Url = find<String>(v, "url");
+                conf.Interval = find<Integer>(v, "interval");
+                conf.Tolerance = find_or<Integer>(v, "tolerance", 0);
+                if(v.contains("lazy"))
+                    conf.Lazy = find_or<bool>(v, "lazy", false);
+                if(v.contains("evaluate-before-use"))
+                    conf.EvaluateBeforeUse = find_or(v, "evaluate-before-use", conf.EvaluateBeforeUse.get());
+                break;
             default:
-                throw toml::syntax_error("Proxy Group has incorrect type, should be one of following:\n  select, url-test, load-balance, fallback, relay, ssid", v.at("type").location());
+                throw serialization_error(format_error("Proxy Group has unsupported type!", v.at("type").location(), "should be one of following: select, url-test, load-balance, fallback, relay, ssid"), v.at("type").location());
             }
-            conf.Timeout = toml::find_or(v, "timeout", 5);
-            conf.Proxies = toml::find_or<StrArray>(v, "rule", {});
-            conf.UsingProvider = toml::find_or<StrArray>(v, "use", {});
+            conf.Timeout = find_or(v, "timeout", 5);
+            conf.Proxies = find_or<StrArray>(v, "rule", {});
+            conf.UsingProvider = find_or<StrArray>(v, "use", {});
             if(conf.Proxies.empty() && conf.UsingProvider.empty())
-                throw toml::syntax_error("Proxy Group must contains at least one of proxy match rule or provider", v.location());
+                throw serialization_error(format_error("Proxy Group must contains at least one of proxy match rule or provider!", v.location(), "here"), v.location());
             if(v.contains("disable-udp"))
-                conf.DisableUdp = toml::find_or(v, "disable-udp", conf.DisableUdp.get());
+                conf.DisableUdp = find_or(v, "disable-udp", conf.DisableUdp.get());
             return conf;
         }
     };
@@ -84,8 +94,8 @@ namespace toml
         static RulesetConfig from_toml(const value& v)
         {
             RulesetConfig conf;
-            conf.Group = toml::find<String>(v, "group");
-            String type = toml::find_or<String>(v, "type", "surge-ruleset");
+            conf.Group = find<String>(v, "group");
+            String type = find_or<String>(v, "type", "surge-ruleset");
             switch(hash_(type))
             {
             /*
@@ -122,10 +132,10 @@ namespace toml
                 conf.Url = type + ":";
                 break;
             default:
-                throw toml::syntax_error("Ruleset has incorrect type, should be one of following:\n  surge-ruleset, quantumultx, clash-domain, clash-ipcidr, clash-classic", v.at("type").location());
+                throw serialization_error(format_error("Ruleset has unsupported type!", v.at("type").location(), "should be one of following: surge-ruleset, quantumultx, clash-domain, clash-ipcidr, clash-classic"), v.at("type").location());
             }
-            conf.Url += toml::find<String>(v, "ruleset");
-            conf.Interval = toml::find_or<Integer>(v, "interval", 86400);
+            conf.Url += find<String>(v, "ruleset");
+            conf.Interval = find_or<Integer>(v, "interval", 86400);
             return conf;
         }
     };
@@ -138,14 +148,14 @@ namespace toml
             RegexMatchConfig conf;
             if(v.contains("script"))
             {
-                conf.Script = toml::find<String>(v, "script");
+                conf.Script = find<String>(v, "script");
                 return conf;
             }
-            conf.Match = toml::find<String>(v, "match");
+            conf.Match = find<String>(v, "match");
             if(v.contains("emoji"))
-                conf.Replace = toml::find<String>(v, "emoji");
+                conf.Replace = find<String>(v, "emoji");
             else
-                conf.Replace = toml::find<String>(v, "replace");
+                conf.Replace = find<String>(v, "replace");
             return conf;
         }
     };
@@ -156,10 +166,10 @@ namespace toml
         static CronTaskConfig from_toml(const value& v)
         {
             CronTaskConfig conf;
-            conf.Name = toml::find<String>(v, "name");
-            conf.CronExp = toml::find<String>(v, "cronexp");
-            conf.Path = toml::find<String>(v, "path");
-            conf.Timeout = toml::find_or<Integer>(v, "timeout", 0);
+            conf.Name = find<String>(v, "name");
+            conf.CronExp = find<String>(v, "cronexp");
+            conf.Path = find<String>(v, "path");
+            conf.Timeout = find_or<Integer>(v, "timeout", 0);
             return conf;
         }
     };
@@ -219,6 +229,9 @@ namespace INIBinding
                     break;
                 case "ssid"_hash:
                     conf.Type = ProxyGroupType::SSID;
+                    break;
+                case "smart"_hash:
+                    conf.Type = ProxyGroupType::Smart;
                     break;
                 default:
                     continue;
