@@ -1,10 +1,11 @@
 #include <string>
 
-#include "../../handler/settings.h"
-#include "../../utils/logger.h"
-#include "../../utils/network.h"
-#include "../../utils/regexp.h"
-#include "../../utils/string.h"
+#include "handler/settings.h"
+#include "utils/logger.h"
+#include "utils/network.h"
+#include "utils/regexp.h"
+#include "utils/string.h"
+#include "utils/rapidjson_extra.h"
 #include "subexport.h"
 
 /// rule type lists
@@ -14,6 +15,7 @@ string_array Surge2RuleTypes = {basic_types, "IP-CIDR6", "USER-AGENT", "URL-REGE
 string_array SurgeRuleTypes = {basic_types, "IP-CIDR6", "USER-AGENT", "URL-REGEX", "AND", "OR", "NOT", "PROCESS-NAME", "IN-PORT", "DEST-PORT", "SRC-IP"};
 string_array QuanXRuleTypes = {basic_types, "USER-AGENT", "HOST", "HOST-SUFFIX", "HOST-KEYWORD"};
 string_array SurfRuleTypes = {basic_types, "IP-CIDR6", "PROCESS-NAME", "IN-PORT", "DEST-PORT", "SRC-IP"};
+string_array SingBoxRuleTypes = {basic_types, "IP-VERSION", "INBOUND", "PROTOCOL", "NETWORK", "GEOSITE", "SRC-GEOIP", "DOMAIN-REGEX", "PROCESS-NAME", "PROCESS-PATH", "PACKAGE-NAME", "PORT", "PORT-RANGE", "SRC-PORT", "SRC-PORT-RANGE", "USER", "USER-ID"};
 
 std::string convertRuleset(const std::string &content, int type)
 {
@@ -28,7 +30,7 @@ std::string convertRuleset(const std::string &content, int type)
 
     if(regFind(content, "^payload:\\r?\\n")) /// Clash
     {
-        output = regReplace(regReplace(content, "payload:\\r?\\n", "", true), "\\s?^\\s*-\\s+('|\"?)(.*)\\1$", "\n$2", true);
+        output = regReplace(regReplace(content, "payload:\\r?\\n", "", true), R"(\s?^\s*-\s+('|"?)(.*)\1$)", "\n$2", true);
         if(type == RULESET_CLASH_CLASSICAL) /// classical type
             return output;
         std::stringstream ss;
@@ -51,8 +53,8 @@ std::string convertRuleset(const std::string &content, int type)
 
             if(!strLine.empty() && (strLine[0] != ';' && strLine[0] != '#' && !(lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')))
             {
-                pos = strLine.find("/");
-                if(pos != strLine.npos) /// ipcidr
+                pos = strLine.find('/');
+                if(pos != std::string::npos) /// ipcidr
                 {
                     if(isIPv4(strLine.substr(0, pos)))
                         output += "IP-CIDR,";
@@ -93,6 +95,32 @@ std::string convertRuleset(const std::string &content, int type)
     }
 }
 
+static std::string transformRuleToCommon(string_view_array &temp, const std::string &input, const std::string &group, bool no_resolve_only = false)
+{
+    temp.clear();
+    std::string strLine;
+    split(temp, input, ',');
+    if(temp.size() < 2)
+    {
+        strLine = temp[0];
+        strLine += ",";
+        strLine += group;
+    }
+    else
+    {
+        strLine = temp[0];
+        strLine += ",";
+        strLine += temp[1];
+        strLine += ",";
+        strLine += group;
+        if(temp.size() > 2 && (!no_resolve_only || temp[2] == "no-resolve"))
+        {
+            strLine += ",";
+            strLine += temp[2];
+        }
+    }
+    return strLine;
+}
 
 void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name)
 {
@@ -106,6 +134,7 @@ void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_
     if(!overwrite_original_rules && base_rule[field_name].IsDefined())
         rules = base_rule[field_name];
 
+    std::vector<std::string_view> temp(4);
     for(RulesetContent &x : ruleset_content_array)
     {
         if(global.maxAllowedRules && total_rules > global.maxAllowedRules)
@@ -122,10 +151,8 @@ void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_
             strLine = retrieved_rules.substr(2);
             if(startsWith(strLine, "FINAL"))
                 strLine.replace(0, 5, "MATCH");
-            strLine += "," + rule_group;
-            if(count_least(strLine, ',', 3))
-                strLine = regReplace(strLine, "^(.*?,.*?)(,.*)(,.*)$", "$1$3$2");
-            allRules.emplace_back(std::move(strLine));
+            strLine = transformRuleToCommon(temp, strLine, rule_group);
+            allRules.emplace_back(strLine);
             total_rules++;
             continue;
         }
@@ -143,18 +170,15 @@ void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_
             lineSize = strLine.size();
             if(!lineSize || strLine[0] == ';' || strLine[0] == '#' || (lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')) //empty lines and comments are ignored
                 continue;
-            if(std::none_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [strLine](std::string type){return startsWith(strLine, type);}))
+            if(std::none_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [strLine](const std::string& type){return startsWith(strLine, type);}))
                 continue;
             if(strFind(strLine, "//"))
             {
                 strLine.erase(strLine.find("//"));
                 strLine = trimWhitespace(strLine);
             }
-            strLine += "," + rule_group;
-            if(count_least(strLine, ',', 3))
-                strLine = regReplace(strLine, "^(.*?,.*?)(,.*)(,.*)$", "$1$3$2");
-            allRules.emplace_back(std::move(strLine));
-            //rules.push_back(strLine);
+            strLine = transformRuleToCommon(temp, strLine, rule_group);
+            allRules.emplace_back(strLine);
         }
     }
 
@@ -177,10 +201,11 @@ std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<RulesetContent>
     if(!overwrite_original_rules && base_rule[field_name].IsDefined())
     {
         for(size_t i = 0; i < base_rule[field_name].size(); i++)
-            output_content += " - " + safe_as<std::string>(base_rule[field_name][i]) + "\n";
+            output_content += "  - " + safe_as<std::string>(base_rule[field_name][i]) + "\n";
     }
     base_rule.remove(field_name);
 
+    string_view_array temp(4);
     for(RulesetContent &x : ruleset_content_array)
     {
         if(global.maxAllowedRules && total_rules > global.maxAllowedRules)
@@ -197,10 +222,8 @@ std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<RulesetContent>
             strLine = retrieved_rules.substr(2);
             if(startsWith(strLine, "FINAL"))
                 strLine.replace(0, 5, "MATCH");
-            strLine += "," + rule_group;
-            if(count_least(strLine, ',', 3))
-                strLine = regReplace(strLine, "^(.*?,.*?)(,.*)(,.*)$", "$1$3$2");
-            output_content += " - " + strLine + "\n";
+            strLine = transformRuleToCommon(temp, strLine, rule_group);
+            output_content += "  - " + strLine + "\n";
             total_rules++;
             continue;
         }
@@ -218,24 +241,22 @@ std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<RulesetContent>
             lineSize = strLine.size();
             if(!lineSize || strLine[0] == ';' || strLine[0] == '#' || (lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')) //empty lines and comments are ignored
                 continue;
-            if(std::none_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [strLine](std::string type){ return startsWith(strLine, type); }))
+            if(std::none_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [strLine](const std::string& type){ return startsWith(strLine, type); }))
                 continue;
             if(strFind(strLine, "//"))
             {
                 strLine.erase(strLine.find("//"));
                 strLine = trimWhitespace(strLine);
             }
-            strLine += "," + rule_group;
-            if(count_least(strLine, ',', 3))
-                strLine = regReplace(strLine, "^(.*?,.*?)(,.*)(,.*)$", "$1$3$2");
-            output_content += " - " + strLine + "\n";
+            strLine = transformRuleToCommon(temp, strLine, rule_group);
+            output_content += "  - " + strLine + "\n";
             total_rules++;
         }
     }
     return output_content;
 }
 
-void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, std::string remote_path_prefix)
+void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, const std::string &remote_path_prefix)
 {
     string_array allRules;
     std::string rule_group, rule_path, rule_path_typed, retrieved_rules, strLine;
@@ -245,34 +266,37 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
     switch(surge_ver) //other version: -3 for Surfboard, -4 for Loon
     {
     case 0:
-        base_rule.SetCurrentSection("RoutingRule"); //Mellow
+        base_rule.set_current_section("RoutingRule"); //Mellow
         break;
     case -1:
-        base_rule.SetCurrentSection("filter_local"); //Quantumult X
+        base_rule.set_current_section("filter_local"); //Quantumult X
         break;
     case -2:
-        base_rule.SetCurrentSection("TCP"); //Quantumult
+        base_rule.set_current_section("TCP"); //Quantumult
         break;
     default:
-        base_rule.SetCurrentSection("Rule");
+        base_rule.set_current_section("Rule");
     }
 
     if(overwrite_original_rules)
     {
-        base_rule.EraseSection();
+        base_rule.erase_section();
         switch(surge_ver)
         {
         case -1:
-            base_rule.EraseSection("filter_remote");
+            base_rule.erase_section("filter_remote");
             break;
         case -4:
-            base_rule.EraseSection("Remote Rule");
+            base_rule.erase_section("Remote Rule");
+            break;
+        default:
             break;
         }
     }
 
     const std::string rule_match_regex = "^(.*?,.*?)(,.*)(,.*)$";
 
+    string_view_array temp(4);
     for(RulesetContent &x : ruleset_content_array)
     {
         if(global.maxAllowedRules && total_rules > global.maxAllowedRules)
@@ -285,21 +309,17 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
             strLine = x.rule_content.get().substr(2);
             if(strLine == "MATCH")
                 strLine = "FINAL";
-            strLine += "," + rule_group;
             if(surge_ver == -1 || surge_ver == -2)
             {
-                if(count_least(strLine, ',', 3) && regReplace(strLine, rule_match_regex, "$2") == ",no-resolve")
-                    strLine = regReplace(strLine, rule_match_regex, "$1$3$2");
-                else
-                    strLine = regReplace(strLine, rule_match_regex, "$1$3");
+                strLine = transformRuleToCommon(temp, strLine, rule_group, true);
             }
             else
             {
-                if(!startsWith(strLine, "AND") && !startsWith(strLine, "OR") && !startsWith(strLine, "NOT") && count_least(strLine, ',', 3))
-                    strLine = regReplace(strLine, rule_match_regex, "$1$3$2");
+                if(!startsWith(strLine, "AND") && !startsWith(strLine, "OR") && !startsWith(strLine, "NOT"))
+                    strLine = transformRuleToCommon(temp, strLine, rule_group);
             }
             strLine = replaceAllDistinct(strLine, ",,", ",");
-            allRules.emplace_back(std::move(strLine));
+            allRules.emplace_back(strLine);
             total_rules++;
             continue;
         }
@@ -308,30 +328,30 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
             if(surge_ver == -1 && x.rule_type == RULESET_QUANX && isLink(rule_path))
             {
                 strLine = rule_path + ", tag=" + rule_group + ", force-policy=" + rule_group + ", enabled=true";
-                base_rule.Set("filter_remote", "{NONAME}", strLine);
+                base_rule.set("filter_remote", "{NONAME}", strLine);
                 continue;
             }
             if(fileExist(rule_path))
             {
-                if(surge_ver > 2 && remote_path_prefix.size())
+                if(surge_ver > 2 && !remote_path_prefix.empty())
                 {
                     strLine = "RULE-SET," + remote_path_prefix + "/getruleset?type=1&url=" + urlSafeBase64Encode(rule_path_typed) + "," + rule_group;
                     if(x.update_interval)
                         strLine += ",update-interval=" + std::to_string(x.update_interval);
-                    allRules.emplace_back(std::move(strLine));
+                    allRules.emplace_back(strLine);
                     continue;
                 }
-                else if(surge_ver == -1 && remote_path_prefix.size())
+                else if(surge_ver == -1 && !remote_path_prefix.empty())
                 {
                     strLine = remote_path_prefix + "/getruleset?type=2&url=" + urlSafeBase64Encode(rule_path_typed) + "&group=" + urlSafeBase64Encode(rule_group);
                     strLine += ", tag=" + rule_group + ", enabled=true";
-                    base_rule.Set("filter_remote", "{NONAME}", strLine);
+                    base_rule.set("filter_remote", "{NONAME}", strLine);
                     continue;
                 }
-                else if(surge_ver == -4 && remote_path_prefix.size())
+                else if(surge_ver == -4 && !remote_path_prefix.empty())
                 {
                     strLine = remote_path_prefix + "/getruleset?type=1&url=" + urlSafeBase64Encode(rule_path_typed) + "," + rule_group;
-                    base_rule.Set("Remote Rule", "{NONAME}", strLine);
+                    base_rule.set("Remote Rule", "{NONAME}", strLine);
                     continue;
                 }
             }
@@ -341,7 +361,7 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
                 {
                     if(x.rule_type != RULESET_SURGE)
                     {
-                        if(remote_path_prefix.size())
+                        if(!remote_path_prefix.empty())
                             strLine = "RULE-SET," + remote_path_prefix + "/getruleset?type=1&url=" + urlSafeBase64Encode(rule_path_typed) + "," + rule_group;
                         else
                             continue;
@@ -352,20 +372,20 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
                     if(x.update_interval)
                         strLine += ",update-interval=" + std::to_string(x.update_interval);
 
-                    allRules.emplace_back(std::move(strLine));
+                    allRules.emplace_back(strLine);
                     continue;
                 }
-                else if(surge_ver == -1 && remote_path_prefix.size())
+                else if(surge_ver == -1 && !remote_path_prefix.empty())
                 {
                     strLine = remote_path_prefix + "/getruleset?type=2&url=" + urlSafeBase64Encode(rule_path_typed) + "&group=" + urlSafeBase64Encode(rule_group);
                     strLine += ", tag=" + rule_group + ", enabled=true";
-                    base_rule.Set("filter_remote", "{NONAME}", strLine);
+                    base_rule.set("filter_remote", "{NONAME}", strLine);
                     continue;
                 }
                 else if(surge_ver == -4)
                 {
                     strLine = rule_path + "," + rule_group;
-                    base_rule.Set("Remote Rule", "{NONAME}", strLine);
+                    base_rule.set("Remote Rule", "{NONAME}", strLine);
                     continue;
                 }
             }
@@ -401,22 +421,22 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
                         continue;
                     [[fallthrough]];
                 case -1:
-                    if(!std::any_of(QuanXRuleTypes.begin(), QuanXRuleTypes.end(), [strLine](std::string type){return startsWith(strLine, type);}))
+                    if(!std::any_of(QuanXRuleTypes.begin(), QuanXRuleTypes.end(), [strLine](const std::string& type){return startsWith(strLine, type);}))
                         continue;
                     break;
                 case -3:
-                    if(!std::any_of(SurfRuleTypes.begin(), SurfRuleTypes.end(), [strLine](std::string type){return startsWith(strLine, type);}))
+                    if(!std::any_of(SurfRuleTypes.begin(), SurfRuleTypes.end(), [strLine](const std::string& type){return startsWith(strLine, type);}))
                         continue;
                     break;
                 default:
                     if(surge_ver > 2)
                     {
-                        if(!std::any_of(SurgeRuleTypes.begin(), SurgeRuleTypes.end(), [strLine](std::string type){return startsWith(strLine, type);}))
+                        if(!std::any_of(SurgeRuleTypes.begin(), SurgeRuleTypes.end(), [strLine](const std::string& type){return startsWith(strLine, type);}))
                             continue;
                     }
                     else
                     {
-                        if(!std::any_of(Surge2RuleTypes.begin(), Surge2RuleTypes.end(), [strLine](std::string type){return startsWith(strLine, type);}))
+                        if(!std::any_of(Surge2RuleTypes.begin(), Surge2RuleTypes.end(), [strLine](const std::string& type){return startsWith(strLine, type);}))
                             continue;
                     }
                 }
@@ -427,22 +447,18 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
                     strLine = trimWhitespace(strLine);
                 }
 
-                strLine += "," + rule_group;
                 if(surge_ver == -1 || surge_ver == -2)
                 {
                     if(startsWith(strLine, "IP-CIDR6"))
                         strLine.replace(0, 8, "IP6-CIDR");
-                    if(count_least(strLine, ',', 3) && regReplace(strLine, rule_match_regex, "$2") == ",no-resolve")
-                        strLine = regReplace(strLine, rule_match_regex, "$1$3$2");
-                    else
-                        strLine = regReplace(strLine, rule_match_regex, "$1$3");
+                    strLine = transformRuleToCommon(temp, strLine, rule_group, true);
                 }
                 else
                 {
-                    if(!startsWith(strLine, "AND") && !startsWith(strLine, "OR") && !startsWith(strLine, "NOT") && count_least(strLine, ',', 3))
-                        strLine = regReplace(strLine, rule_match_regex, "$1$3$2");
+                    if(!startsWith(strLine, "AND") && !startsWith(strLine, "OR") && !startsWith(strLine, "NOT"))
+                        strLine = transformRuleToCommon(temp, strLine, rule_group);
                 }
-                allRules.emplace_back(std::move(strLine));
+                allRules.emplace_back(strLine);
                 total_rules++;
             }
         }
@@ -450,6 +466,141 @@ void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_c
 
     for(std::string &x : allRules)
     {
-        base_rule.Set("{NONAME}", x);
+        base_rule.set("{NONAME}", x);
     }
+}
+
+static rapidjson::Value transformRuleToSingBox(std::vector<std::string_view> &args, const std::string& rule, const std::string &group, rapidjson::MemoryPoolAllocator<>& allocator)
+{
+    args.clear();
+    split(args, rule, ',');
+    if (args.size() < 2) return rapidjson::Value(rapidjson::kObjectType);
+    auto type = toLower(std::string(args[0]));
+    auto value = toLower(std::string(args[1]));
+//    std::string_view option;
+//    if (args.size() >= 3) option = args[2];
+
+    rapidjson::Value rule_obj(rapidjson::kObjectType);
+    type = replaceAllDistinct(type, "-", "_");
+    type = replaceAllDistinct(type, "ip_cidr6", "ip_cidr");
+    type = replaceAllDistinct(type, "src_", "source_");
+    if (type == "match" || type == "final")
+    {
+        rule_obj.AddMember("outbound", rapidjson::Value(value.data(), value.size(), allocator), allocator);
+    }
+    else
+    {
+        rule_obj.AddMember(rapidjson::Value(type.c_str(), allocator), rapidjson::Value(value.data(), value.size(), allocator), allocator);
+        rule_obj.AddMember("outbound", rapidjson::Value(group.c_str(), allocator), allocator);
+    }
+    return rule_obj;
+}
+
+static void appendSingBoxRule(std::vector<std::string_view> &args, rapidjson::Value &rules, const std::string& rule, rapidjson::MemoryPoolAllocator<>& allocator)
+{
+    using namespace rapidjson_ext;
+    args.clear();
+    split(args, rule, ',');
+    if (args.size() < 2) return;
+    auto type = args[0];
+//    std::string_view option;
+//    if (args.size() >= 3) option = args[2];
+
+    if (none_of(SingBoxRuleTypes, [&](const std::string& t){ return type == t; }))
+        return;
+
+    auto realType = toLower(std::string(type));
+    auto value = toLower(std::string(args[1]));
+    realType = replaceAllDistinct(realType, "-", "_");
+    realType = replaceAllDistinct(realType, "ip_cidr6", "ip_cidr");
+
+    rules | AppendToArray(realType.c_str(), rapidjson::Value(value.c_str(), value.size(), allocator), allocator);
+}
+
+void rulesetToSingBox(rapidjson::Document &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules)
+{
+    using namespace rapidjson_ext;
+    std::string rule_group, retrieved_rules, strLine, final;
+    std::stringstream strStrm;
+    size_t total_rules = 0;
+    auto &allocator = base_rule.GetAllocator();
+
+    rapidjson::Value rules(rapidjson::kArrayType);
+    if (!overwrite_original_rules)
+    {
+        if (base_rule.HasMember("route") && base_rule["route"].HasMember("rules") && base_rule["route"]["rules"].IsArray())
+            rules.Swap(base_rule["route"]["rules"]);
+    }
+
+    if (global.singBoxAddClashModes)
+    {
+        auto global_object = buildObject(allocator, "clash_mode", "Global", "outbound", "GLOBAL");
+        auto direct_object = buildObject(allocator, "clash_mode", "Direct", "outbound", "DIRECT");
+        rules.PushBack(global_object, allocator);
+        rules.PushBack(direct_object, allocator);
+    }
+
+    auto dns_object = buildObject(allocator, "protocol", "dns", "outbound", "dns-out");
+    rules.PushBack(dns_object, allocator);
+
+    std::vector<std::string_view> temp(4);
+    for(RulesetContent &x : ruleset_content_array)
+    {
+        if(global.maxAllowedRules && total_rules > global.maxAllowedRules)
+            break;
+        rule_group = x.rule_group;
+        retrieved_rules = x.rule_content.get();
+        if(retrieved_rules.empty())
+        {
+            writeLog(0, "Failed to fetch ruleset or ruleset is empty: '" + x.rule_path + "'!", LOG_LEVEL_WARNING);
+            continue;
+        }
+        if(startsWith(retrieved_rules, "[]"))
+        {
+            strLine = retrieved_rules.substr(2);
+            if(startsWith(strLine, "FINAL") || startsWith(strLine, "MATCH"))
+            {
+                final = rule_group;
+                continue;
+            }
+            rules.PushBack(transformRuleToSingBox(temp, strLine, rule_group, allocator), allocator);
+            total_rules++;
+            continue;
+        }
+        retrieved_rules = convertRuleset(retrieved_rules, x.rule_type);
+        char delimiter = getLineBreak(retrieved_rules);
+
+        strStrm.clear();
+        strStrm<<retrieved_rules;
+
+        std::string::size_type lineSize;
+        rapidjson::Value rule(rapidjson::kObjectType);
+
+        while(getline(strStrm, strLine, delimiter))
+        {
+            if(global.maxAllowedRules && total_rules > global.maxAllowedRules)
+                break;
+            strLine = trimWhitespace(strLine, true, true); //remove whitespaces
+            lineSize = strLine.size();
+            if(!lineSize || strLine[0] == ';' || strLine[0] == '#' || (lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')) //empty lines and comments are ignored
+                continue;
+            if(strFind(strLine, "//"))
+            {
+                strLine.erase(strLine.find("//"));
+                strLine = trimWhitespace(strLine);
+            }
+            appendSingBoxRule(temp, rule, strLine, allocator);
+        }
+        if (rule.ObjectEmpty()) continue;
+        rule.AddMember("outbound", rapidjson::Value(rule_group.c_str(), allocator), allocator);
+        rules.PushBack(rule, allocator);
+    }
+
+    if (!base_rule.HasMember("route"))
+        base_rule.AddMember("route", rapidjson::Value(rapidjson::kObjectType), allocator);
+
+    auto finalValue = rapidjson::Value(final.c_str(), allocator);
+    base_rule["route"]
+    | AddMemberOrReplace("rules", rules, allocator)
+    | AddMemberOrReplace("final", finalValue, allocator);
 }
