@@ -6,8 +6,9 @@
 #include <future>
 
 #include <yaml-cpp/yaml.h>
+#include <rapidjson/document.h>
 
-#include "../../utils/ini_reader/ini_reader.h"
+#include "utils/ini_reader/ini_reader.h"
 
 enum ruleset_type
 {
@@ -31,6 +32,7 @@ struct RulesetContent
 std::string convertRuleset(const std::string &content, int type);
 void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
 std::string rulesetToClashStr(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
-void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, std::string remote_path_prefix);
+void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, const std::string& remote_path_prefix);
+void rulesetToSingBox(rapidjson::Document &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules);
 
 #endif // RULECONVERT_H_INCLUDED
