@@ -4,22 +4,22 @@
 #include <cmath>
 #include <climits>
 
-#include "../../config/regmatch.h"
-#include "../../generator/config/subexport.h"
-#include "../../generator/template/templates.h"
-#include "../../handler/settings.h"
-#include "../../parser/config/proxy.h"
-#include "../../script/script_quickjs.h"
-#include "../../utils/bitwise.h"
-#include "../../utils/file_extra.h"
-#include "../../utils/ini_reader/ini_reader.h"
-#include "../../utils/logger.h"
-#include "../../utils/network.h"
-#include "../../utils/rapidjson_extra.h"
-#include "../../utils/regexp.h"
-#include "../../utils/stl_extra.h"
-#include "../../utils/urlencode.h"
-#include "../../utils/yamlcpp_extra.h"
+#include "config/regmatch.h"
+#include "generator/config/subexport.h"
+#include "generator/template/templates.h"
+#include "handler/settings.h"
+#include "parser/config/proxy.h"
+#include "script/script_quickjs.h"
+#include "utils/bitwise.h"
+#include "utils/file_extra.h"
+#include "utils/ini_reader/ini_reader.h"
+#include "utils/logger.h"
+#include "utils/network.h"
+#include "utils/rapidjson_extra.h"
+#include "utils/regexp.h"
+#include "utils/stl_extra.h"
+#include "utils/urlencode.h"
+#include "utils/yamlcpp_extra.h"
 #include "nodemanip.h"
 #include "ruleconvert.h"
 
@@ -116,7 +116,19 @@ bool applyMatcher(const std::string &rule, std::string &real_rule, const Proxy &
     std::string target, ret_real_rule;
     static const std::string groupid_regex = R"(^!!(?:GROUPID|INSERT)=([\d\-+!,]+)(?:!!(.*))?$)", group_regex = R"(^!!(?:GROUP)=(.+?)(?:!!(.*))?$)";
     static const std::string type_regex = R"(^!!(?:TYPE)=(.+?)(?:!!(.*))?$)", port_regex = R"(^!!(?:PORT)=(.+?)(?:!!(.*))?$)", server_regex = R"(^!!(?:SERVER)=(.+?)(?:!!(.*))?$)";
-    static const string_array types = {"", "SS", "SSR", "VMESS", "TROJAN", "SNELL", "HTTP", "HTTPS", "SOCKS5"};
+    static const std::map<ProxyType, const char *> types = {
+        {ProxyType::Shadowsocks,  "SS"},
+        {ProxyType::ShadowsocksR, "SSR"},
+        {ProxyType::VMess,        "VMESS"},
+        {ProxyType::Trojan,       "TROJAN"},
+        {ProxyType::Snell,        "SNELL"},
+        {ProxyType::HTTP,         "HTTP"},
+        {ProxyType::HTTPS,        "HTTPS"},
+        {ProxyType::SOCKS5,       "SOCKS5"},
+        {ProxyType::WireGuard,    "WIREGUARD"},
+        {ProxyType::Hysteria,     "HYSTERIA"},
+        {ProxyType::Hysteria2,    "HYSTERIA2"}
+    };
     if(startsWith(rule, "!!GROUP="))
     {
         regGetMatch(rule, group_regex, 3, 0, &target, &ret_real_rule);
@@ -134,9 +146,9 @@ bool applyMatcher(const std::string &rule, std::string &real_rule, const Proxy &
     {
         regGetMatch(rule, type_regex, 3, 0, &target, &ret_real_rule);
         real_rule = ret_real_rule;
-        if(node.Type == ProxyType::Unknow)
+        if(node.Type == ProxyType::Unknown)
             return false;
-        return regMatch(types[node.Type], target);
+        return regMatch(types.at(node.Type), target);
     }
     else if(startsWith(rule, "!!PORT="))
     {
@@ -155,24 +167,29 @@ bool applyMatcher(const std::string &rule, std::string &real_rule, const Proxy &
     return true;
 }
 
-void processRemark(std::string &oldremark, std::string &newremark, string_array &remarks_list, bool proc_comma = true)
+void processRemark(std::string &remark, const string_array &remarks_list, bool proc_comma = true)
 {
+    // Replace every '=' with '-' in the remark string to avoid parse errors from the clients.
+    //     Surge is tested to yield an error when handling '=' in the remark string, 
+    //     not sure if other clients have the same problem.
+    std::replace(remark.begin(), remark.end(), '=', '-');
+
     if(proc_comma)
     {
-        if(oldremark.find(',') != oldremark.npos)
+        if(remark.find(',') != std::string::npos)
         {
-            oldremark.insert(0, "\"");
-            oldremark.append("\"");
+            remark.insert(0, "\"");
+            remark.append("\"");
         }
     }
-    newremark = oldremark;
+    std::string tempRemark = remark;
     int cnt = 2;
-    while(std::find(remarks_list.begin(), remarks_list.end(), newremark) != remarks_list.end())
+    while(std::find(remarks_list.cbegin(), remarks_list.cend(), tempRemark) != remarks_list.cend())
     {
-        newremark = oldremark + " " + std::to_string(cnt);
+        tempRemark = remark + " " + std::to_string(cnt);
         cnt++;
     }
-    oldremark = newremark;
+    remark = tempRemark;
 }
 
 void groupGenerate(const std::string &rule, std::vector<Proxy> &nodelist, string_array &filtered_nodelist, bool add_direct, extra_settings &ext)
@@ -213,41 +230,53 @@ void groupGenerate(const std::string &rule, std::vector<Proxy> &nodelist, string
 
 void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGroupConfigs &extra_proxy_group, bool clashR, extra_settings &ext)
 {
-    YAML::Node proxies, singleproxy, singlegroup, original_groups;
+    YAML::Node proxies, original_groups;
     std::vector<Proxy> nodelist;
-    string_array remarks_list, filtered_nodelist;
+    string_array remarks_list;
     /// proxies style
-    bool block = false, compact = false;
+    bool proxy_block = false, proxy_compact = false, group_block = false, group_compact = false;
     switch(hash_(ext.clash_proxies_style))
     {
     case "block"_hash:
-        block = true;
+        proxy_block = true;
         break;
     default:
     case "flow"_hash:
         break;
     case "compact"_hash:
-        compact = true;
+        proxy_compact = true;
         break;
+    }
+    switch(hash_(ext.clash_proxy_groups_style))
+    {
+        case "block"_hash:
+            group_block = true;
+            break;
+        default:
+        case "flow"_hash:
+            break;
+        case "compact"_hash:
+            group_compact = true;
+            break;
     }
 
     for(Proxy &x : nodes)
     {
-        singleproxy.reset();
+        YAML::Node singleproxy;
 
         std::string type = getProxyTypeName(x.Type);
-        std::string remark, pluginopts = replaceAllDistinct(x.PluginOption, ";", "&");
+        std::string pluginopts = replaceAllDistinct(x.PluginOption, ";", "&");
         if(ext.append_proxy_type)
             x.Remark = "[" + type + "] " + x.Remark;
 
-        processRemark(x.Remark, remark, remarks_list, false);
+        processRemark(x.Remark, remarks_list, false);
 
-        tribool udp = ext.udp;
-        tribool scv = ext.skip_cert_verify;
+        tribool udp = ext.udp, tfo = ext.tfo, scv = ext.skip_cert_verify;
         udp.define(x.UDP);
+        tfo.define(x.TCPFastOpen);
         scv.define(x.AllowInsecure);
 
-        singleproxy["name"] = remark;
+        singleproxy["name"] = x.Remark;
         singleproxy["server"] = x.Hostname;
         singleproxy["port"] = x.Port;
 
@@ -424,6 +453,8 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
             }
             break;
         case ProxyType::Snell:
+            if (x.SnellVersion >= 4)
+                continue;
             singleproxy["type"] = "snell";
             singleproxy["psk"] = x.Password;
             if(x.SnellVersion != 0)
@@ -437,22 +468,111 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
             if(std::all_of(x.Password.begin(), x.Password.end(), ::isdigit) && !x.Password.empty())
                 singleproxy["password"].SetTag("str");
             break;
+        case ProxyType::WireGuard:
+            singleproxy["type"] = "wireguard";
+            singleproxy["public-key"] = x.PublicKey;
+            singleproxy["private-key"] = x.PrivateKey;
+            singleproxy["ip"] = x.SelfIP;
+            if(!x.SelfIPv6.empty())
+                singleproxy["ipv6"] = x.SelfIPv6;
+            if(!x.PreSharedKey.empty())
+                singleproxy["preshared-key"] = x.PreSharedKey;
+            if(!x.DnsServers.empty())
+                singleproxy["dns"] = x.DnsServers;
+            if(x.Mtu > 0)
+                singleproxy["mtu"] = x.Mtu;
+            break;
+        case ProxyType::Hysteria:
+            singleproxy["type"] = "hysteria";
+            if (!x.Ports.empty())
+                singleproxy["ports"] = x.Ports;
+            if (!x.Protocol.empty())
+                singleproxy["protocol"] = x.Protocol;
+            if (!x.OBFSParam.empty())
+                singleproxy["obfs-protocol"] = x.OBFSParam;
+            if (!x.Up.empty())
+                singleproxy["up"] = x.Up;
+            if (x.UpSpeed)
+                singleproxy["up-speed"] = x.UpSpeed;
+            if (!x.Down.empty())
+                singleproxy["down"] = x.Down;
+            if (x.DownSpeed)
+                singleproxy["down-speed"] = x.DownSpeed;
+            if (!x.AuthStr.empty())
+            {
+                singleproxy["auth-str"] = x.AuthStr;
+                singleproxy["auth"] = base64Encode(x.AuthStr);
+            }
+            if (!x.OBFS.empty())
+                singleproxy["obfs"] = x.OBFS;
+            if (!x.SNI.empty())
+                singleproxy["sni"] = x.SNI;
+            if (!scv.is_undef())
+                singleproxy["skip-cert-verify"] = scv.get();
+            if (!x.Fingerprint.empty())
+                singleproxy["fingerprint"] = x.Fingerprint;
+            if (!x.Alpn.empty())
+                singleproxy["alpn"] = x.Alpn;
+            if (!x.Ca.empty())
+                singleproxy["ca"] = x.Ca;
+            if (!x.CaStr.empty())
+                singleproxy["ca-str"] = x.CaStr;
+            if (x.RecvWindowConn)
+                singleproxy["recv-window-conn"] = x.RecvWindowConn;
+            if (x.RecvWindow)
+                singleproxy["recv-window"] = x.RecvWindow;
+            if (!x.DisableMtuDiscovery.is_undef())
+                singleproxy["disable-mtu-discovery"] = x.DisableMtuDiscovery.get();
+            if (!x.TCPFastOpen.is_undef())
+                singleproxy["fast-open"] = x.TCPFastOpen.get();
+            if (x.HopInterval)
+                singleproxy["hop-interval"] = x.HopInterval;
+            break;
+        case ProxyType::Hysteria2:
+            singleproxy["type"] = "hysteria2";
+            if (!x.Up.empty())
+                singleproxy["up"] = x.UpSpeed;
+            if (!x.Down.empty())
+                singleproxy["down"] = x.DownSpeed;
+            if (!x.Password.empty())
+                singleproxy["password"] = x.Password;
+            if (!x.OBFS.empty())
+                singleproxy["obfs"] = x.OBFS;
+            if (!x.OBFSParam.empty())
+                singleproxy["obfs-password"] = x.OBFSParam;
+            if (!x.SNI.empty())
+                singleproxy["sni"] = x.SNI;
+            if (!scv.is_undef())
+                singleproxy["skip-cert-verify"] = scv.get();
+            if (!x.Alpn.empty())
+                singleproxy["alpn"] = x.Alpn;
+            if (!x.Ca.empty())
+                singleproxy["ca"] = x.Ca;
+            if (!x.CaStr.empty())
+                singleproxy["ca-str"] = x.CaStr;
+            if (x.CWND)
+                singleproxy["cwnd"] = x.CWND;
+            break;
         default:
             continue;
         }
 
-        if(udp)
+        // UDP is not supported yet in clash using snell
+        // sees in https://dreamacro.github.io/clash/configuration/outbound.html#snell
+        if(udp && x.Type != ProxyType::Snell)
             singleproxy["udp"] = true;
-        if(block)
+        if(!tfo.is_undef())
+            singleproxy["tfo"] = tfo.get();
+        if(proxy_block)
             singleproxy.SetStyle(YAML::EmitterStyle::Block);
         else
             singleproxy.SetStyle(YAML::EmitterStyle::Flow);
         proxies.push_back(singleproxy);
-        remarks_list.emplace_back(std::move(remark));
+        remarks_list.emplace_back(x.Remark);
         nodelist.emplace_back(x);
     }
 
-    if(compact)
+    if(proxy_compact)
         proxies.SetStyle(YAML::EmitterStyle::Flow);
 
     if(ext.nodelist)
@@ -471,11 +591,14 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
 
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
-        singlegroup.reset();
-        eraseElements(filtered_nodelist);
+        YAML::Node singlegroup;
+        string_array filtered_nodelist;
 
         singlegroup["name"] = x.Name;
-        singlegroup["type"] = x.TypeStr();
+        if (x.Type == ProxyGroupType::Smart)
+            singlegroup["type"] = "url-test";
+        else
+            singlegroup["type"] = x.TypeStr();
 
         switch(x.Type)
         {
@@ -484,6 +607,8 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
             break;
         case ProxyGroupType::LoadBalance:
             singlegroup["strategy"] = x.StrategyStr();
+            [[fallthrough]];
+        case ProxyGroupType::Smart:
             [[fallthrough]];
         case ProxyGroupType::URLTest:
             if(!x.Lazy.is_undef())
@@ -514,14 +639,17 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
         }
         if(!filtered_nodelist.empty())
             singlegroup["proxies"] = filtered_nodelist;
-        //singlegroup.SetStyle(YAML::EmitterStyle::Flow);
+        if(group_block)
+            singlegroup.SetStyle(YAML::EmitterStyle::Block);
+        else
+            singlegroup.SetStyle(YAML::EmitterStyle::Flow);
 
         bool replace_flag = false;
-        for(unsigned int i = 0; i < original_groups.size(); i++)
+        for(auto && original_group : original_groups)
         {
-            if(original_groups[i]["name"].as<std::string>() == x.Name)
+            if(original_group["name"].as<std::string>() == x.Name)
             {
-                original_groups[i] = singlegroup;
+                original_group.reset(singlegroup);
                 replace_flag = true;
                 break;
             }
@@ -529,6 +657,9 @@ void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGr
         if(!replace_flag)
             original_groups.push_back(singlegroup);
     }
+
+    if(group_compact)
+        original_groups.SetStyle(YAML::EmitterStyle::Flow);
 
     if(ext.clash_new_field_name)
         yamlnode["proxy-groups"] = original_groups;
@@ -547,7 +678,7 @@ std::string proxyToClash(std::vector<Proxy> &nodes, const std::string &base_conf
     catch (std::exception &e)
     {
         writeLog(0, std::string("Clash base loader failed with error: ") + e.what(), LOG_LEVEL_ERROR);
-        return std::string();
+        return "";
     }
 
     proxyToClash(nodes, yamlnode, extra_proxy_group, clashR, ext);
@@ -586,64 +717,78 @@ std::string proxyToClash(std::vector<Proxy> &nodes, const std::string &base_conf
     return output_content;
 }
 
+// peer = (public-key = bmXOC+F1FxEMF9dyiK2H5/1SUtzH0JuVo51h2wPfgyo=, allowed-ips = "0.0.0.0/0, ::/0", endpoint = engage.cloudflareclient.com:2408, client-id = 139/184/125),(public-key = bmXOC+F1FxEMF9dyiK2H5/1SUtzH0JuVo51h2wPfgyo=, endpoint = engage.cloudflareclient.com:2408)
+std::string generatePeer(Proxy &node, bool client_id_as_reserved = false)
+{
+    std::string result;
+    result += "public-key = " + node.PublicKey;
+    result += ", endpoint = " + node.Hostname + ":" + std::to_string(node.Port);
+    if(!node.AllowedIPs.empty())
+        result += ", allowed-ips = \"" + node.AllowedIPs + "\"";
+    if(!node.ClientId.empty())
+    {
+        if(client_id_as_reserved)
+            result += ", reserved = [" + node.ClientId + "]";
+        else
+            result += ", client-id = " + node.ClientId;
+    }
+    return result;
+}
+
 std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, int surge_ver, extra_settings &ext)
 {
     INIReader ini;
-    std::string proxy;
     std::string output_nodelist;
-    tribool udp, tfo, scv, tls13;
     std::vector<Proxy> nodelist;
     unsigned short local_port = 1080;
-
-    string_array remarks_list, filtered_nodelist, args;
+    string_array remarks_list;
 
     ini.store_any_line = true;
     // filter out sections that requires direct-save
-    ini.AddDirectSaveSection("General");
-    ini.AddDirectSaveSection("Replica");
-    ini.AddDirectSaveSection("Rule");
-    ini.AddDirectSaveSection("MITM");
-    ini.AddDirectSaveSection("Script");
-    ini.AddDirectSaveSection("Host");
-    ini.AddDirectSaveSection("URL Rewrite");
-    ini.AddDirectSaveSection("Header Rewrite");
-    if(ini.Parse(base_conf) != 0 && !ext.nodelist)
+    ini.add_direct_save_section("General");
+    ini.add_direct_save_section("Replica");
+    ini.add_direct_save_section("Rule");
+    ini.add_direct_save_section("MITM");
+    ini.add_direct_save_section("Script");
+    ini.add_direct_save_section("Host");
+    ini.add_direct_save_section("URL Rewrite");
+    ini.add_direct_save_section("Header Rewrite");
+    if(ini.parse(base_conf) != 0 && !ext.nodelist)
     {
-        writeLog(0, "Surge base loader failed with error: " + ini.GetLastError(), LOG_LEVEL_ERROR);
-        return std::string();
+        writeLog(0, "Surge base loader failed with error: " + ini.get_last_error(), LOG_LEVEL_ERROR);
+        return "";
     }
 
-    ini.SetCurrentSection("Proxy");
-    ini.EraseSection();
-    ini.Set("{NONAME}", "DIRECT = direct");
+    ini.set_current_section("Proxy");
+    ini.erase_section();
+    ini.set("{NONAME}", "DIRECT = direct");
 
     for(Proxy &x : nodes)
     {
-        std::string remark;
         if(ext.append_proxy_type)
         {
             std::string type = getProxyTypeName(x.Type);
             x.Remark = "[" + type + "] " + x.Remark;
         }
 
-        processRemark(x.Remark, remark, remarks_list);
+        processRemark(x.Remark, remarks_list);
 
-        std::string &hostname = x.Hostname, &username = x.Username, &password = x.Password, &method = x.EncryptMethod, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &edge = x.Edge, &path = x.Path, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam, &plugin = x.Plugin, &pluginopts = x.PluginOption;
+        std::string &hostname = x.Hostname, &username = x.Username, &password = x.Password, &method = x.EncryptMethod, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &edge = x.Edge, &path = x.Path, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam, &plugin = x.Plugin, &pluginopts = x.PluginOption, &underlying_proxy = x.UnderlyingProxy;
         std::string port = std::to_string(x.Port);
         bool &tlssecure = x.TLSSecure;
 
-        udp = ext.udp;
-        tfo = ext.tfo;
-        scv = ext.skip_cert_verify;
-        tls13 = ext.tls13;
+        tribool udp = ext.udp, tfo = ext.tfo, scv = ext.skip_cert_verify, tls13 = ext.tls13;
         udp.define(x.UDP);
         tfo.define(x.TCPFastOpen);
         scv.define(x.AllowInsecure);
         tls13.define(x.TLS13);
 
-        proxy.clear();
+        std::string proxy, section, real_section;
+        string_array args, headers;
 
-        switch(x.Type)
+        std::stringstream ss;
+
+        switch (x.Type)
         {
         case ProxyType::Shadowsocks:
             if(surge_ver >= 3 || surge_ver == -3)
@@ -679,9 +824,16 @@ std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf
             case "tcp"_hash:
                 break;
             case "ws"_hash:
-                proxy += ", ws=true, ws-path=" + path + ", sni=" + host + ", ws-headers=Host:" + host;
+                if(host.empty())
+                    proxy += ", ws=true, ws-path=" + path + ", sni=" + hostname;
+                else
+                    proxy += ", ws=true, ws-path=" + path + ", sni=" + host;
+                if(!host.empty())
+                    headers.push_back("Host:" + host);
                 if(!edge.empty())
-                    proxy += "|Edge:" + edge;
+                    headers.push_back("Edge:" + edge);
+                if(!headers.empty())
+                    proxy += ", ws-headers=" + join(headers, "|");
                 break;
             default:
                 continue;
@@ -754,7 +906,50 @@ std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf
         case ProxyType::Snell:
             proxy = "snell, " + hostname + ", " + port + ", psk=" + password;
             if(!obfs.empty())
-                proxy += ", obfs=" + obfs + ", obfs-host=" + host;
+            {
+                proxy += ", obfs=" + obfs;
+                if(!host.empty())
+                    proxy += ", obfs-host=" + host;
+            }
+            if(x.SnellVersion != 0)
+                proxy += ", version=" + std::to_string(x.SnellVersion);
+            break;
+        case ProxyType::WireGuard:
+            if(surge_ver < 4 && surge_ver != -3)
+                continue;
+            ss << std::hex << hash_(x.Remark);
+            section = ss.str().substr(0, 5);
+            real_section = "WireGuard " + section;
+            proxy = "wireguard, section-name=" + section;
+            if(!x.TestUrl.empty())
+                proxy += ", test-url=" + x.TestUrl;
+            ini.set(real_section, "private-key", x.PrivateKey);
+            ini.set(real_section, "self-ip", x.SelfIP);
+            if(!x.SelfIPv6.empty())
+                ini.set(real_section, "self-ip-v6", x.SelfIPv6);
+            if(!x.PreSharedKey.empty())
+                ini.set(real_section, "preshared-key", x.PreSharedKey);
+            if(!x.DnsServers.empty())
+                ini.set(real_section, "dns-server", join(x.DnsServers, ","));
+            if(x.Mtu > 0)
+                ini.set(real_section, "mtu", std::to_string(x.Mtu));
+            if(x.KeepAlive > 0)
+                ini.set(real_section, "keepalive", std::to_string(x.KeepAlive));
+            ini.set(real_section, "peer", "(" + generatePeer(x) + ")");
+            break;
+        case ProxyType::Hysteria2:
+            if(surge_ver < 4)
+                continue;
+            proxy = "hysteria, " + hostname + ", " + port + ", password=" + password;
+            if(x.DownSpeed)
+                proxy += ", download-bandwidth=" + x.DownSpeed;
+            
+            if(!scv.is_undef())
+                proxy += ",skip-cert-verify=" + std::string(scv.get() ? "true" : "false");
+            if(!x.Fingerprint.empty())
+                proxy += ",server-cert-fingerprint-sha256=" + x.Fingerprint;
+            if(!x.SNI.empty())
+                proxy += ",sni=" + x.SNI;
             break;
         default:
             continue;
@@ -765,29 +960,33 @@ std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf
         if(!udp.is_undef())
             proxy += ", udp-relay=" + udp.get_str();
 
-        if(ext.nodelist)
-            output_nodelist += remark + " = " + proxy + "\n";
+        if (underlying_proxy != "")
+            proxy += ", underlying-proxy=" + underlying_proxy;
+
+        if (ext.nodelist)
+            output_nodelist += x.Remark + " = " + proxy + "\n";
         else
         {
-            ini.Set("{NONAME}", remark + " = " + proxy);
+            ini.set("{NONAME}", x.Remark + " = " + proxy);
             nodelist.emplace_back(x);
         }
-        remarks_list.emplace_back(std::move(remark));
+        remarks_list.emplace_back(x.Remark);
     }
 
     if(ext.nodelist)
         return output_nodelist;
 
-    ini.SetCurrentSection("Proxy Group");
-    ini.EraseSection();
+    ini.set_current_section("Proxy Group");
+    ini.erase_section();
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
-        eraseElements(filtered_nodelist);
-        proxy.clear();
+        string_array filtered_nodelist;
+        std::string group;
 
         switch(x.Type)
         {
         case ProxyGroupType::Select:
+        case ProxyGroupType::Smart:
         case ProxyGroupType::URLTest:
         case ProxyGroupType::Fallback:
             break;
@@ -796,9 +995,9 @@ std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf
                 continue;
             break;
         case ProxyGroupType::SSID:
-            proxy = x.TypeStr() + ",default=" + x.Proxies[0] + ",";
-            proxy += join(x.Proxies.begin() + 1, x.Proxies.end(), ",");
-            ini.Set("{NONAME}", x.Name + " = " + proxy); //insert order
+            group = x.TypeStr() + ",default=" + x.Proxies[0] + ",";
+                group += join(x.Proxies.begin() + 1, x.Proxies.end(), ",");
+                ini.set("{NONAME}", x.Name + " = " + group); //insert order
             continue;
         default:
             continue;
@@ -812,54 +1011,50 @@ std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf
 
         if(filtered_nodelist.size() == 1)
         {
-            proxy = toLower(filtered_nodelist[0]);
-            switch(hash_(proxy))
+            group = toLower(filtered_nodelist[0]);
+            switch(hash_(group))
             {
             case "direct"_hash:
             case "reject"_hash:
             case "reject-tinygif"_hash:
-                ini.Set("Proxy", "{NONAME}", x.Name + " = " + proxy);
+                ini.set("Proxy", "{NONAME}", x.Name + " = " + group);
                 continue;
             }
         }
 
-        proxy = x.TypeStr() + ",";
-        proxy += join(filtered_nodelist, ",");
+        group = x.TypeStr() + ",";
+        group += join(filtered_nodelist, ",");
         if(x.Type == ProxyGroupType::URLTest || x.Type == ProxyGroupType::Fallback || x.Type == ProxyGroupType::LoadBalance)
         {
-            proxy += ",url=" + x.Url + ",interval=" + std::to_string(x.Interval);
+            group += ",url=" + x.Url + ",interval=" + std::to_string(x.Interval);
             if(x.Tolerance > 0)
-                proxy += ",tolerance=" + std::to_string(x.Tolerance);
+                group += ",tolerance=" + std::to_string(x.Tolerance);
             if(x.Timeout > 0)
-                proxy += ",timeout=" + std::to_string(x.Timeout);
+                group += ",timeout=" + std::to_string(x.Timeout);
             if(!x.Persistent.is_undef())
-                proxy += ",persistent=" + x.Persistent.get_str();
+                group += ",persistent=" + x.Persistent.get_str();
             if(!x.EvaluateBeforeUse.is_undef())
-                proxy += ",evaluate-before-use=" + x.EvaluateBeforeUse.get_str();
+                group += ",evaluate-before-use=" + x.EvaluateBeforeUse.get_str();
         }
 
-        ini.Set("{NONAME}", x.Name + " = " + proxy); //insert order
+        ini.set("{NONAME}", x.Name + " = " + group); //insert order
     }
 
     if(ext.enable_rule_generator)
         rulesetToSurge(ini, ruleset_content_array, surge_ver, ext.overwrite_original_rules, ext.managed_config_prefix);
 
-    return ini.ToString();
+    return ini.to_string();
 }
 
 std::string proxyToSingle(std::vector<Proxy> &nodes, int types, extra_settings &ext)
 {
     /// types: SS=1 SSR=2 VMess=4 Trojan=8
-    std::string remark, hostname, port, password, method;
-    std::string plugin, pluginopts;
-    std::string protocol, protoparam, obfs, obfsparam;
-    std::string id, aid, transproto, faketype, host, path, quicsecure, quicsecret;
     std::string proxyStr, allLinks;
     bool ss = GETBIT(types, 1), ssr = GETBIT(types, 2), vmess = GETBIT(types, 3), trojan = GETBIT(types, 4);
 
     for(Proxy &x : nodes)
     {
-        remark = x.Remark;
+        std::string remark = x.Remark;
         std::string &hostname = x.Hostname, &password = x.Password, &method = x.EncryptMethod, &plugin = x.Plugin, &pluginopts = x.PluginOption, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &path = x.Path, &faketype = x.FakeType;
         bool &tlssecure = x.TLSSecure;
         std::string port = std::to_string(x.Port);
@@ -934,42 +1129,23 @@ std::string proxyToSingle(std::vector<Proxy> &nodes, int types, extra_settings &
 
 std::string proxyToSSSub(std::string base_conf, std::vector<Proxy> &nodes, extra_settings &ext)
 {
-    rapidjson::Document json, base;
-    std::string remark, hostname, password, method;
-    std::string plugin, pluginopts;
-    std::string protocol, obfs;
-    std::string output_content;
+    using namespace rapidjson_ext;
+    rapidjson::Document base;
 
-    rapidjson::Document::AllocatorType &alloc = json.GetAllocator();
-    json.SetObject();
-    json.AddMember("remarks", "", alloc);
-    json.AddMember("server", "", alloc);
-    json.AddMember("server_port", 0, alloc);
-    json.AddMember("method", "", alloc);
-    json.AddMember("password", "", alloc);
-    json.AddMember("plugin", "", alloc);
-    json.AddMember("plugin_opts", "", alloc);
+    auto &alloc = base.GetAllocator();
 
     base_conf = trimWhitespace(base_conf);
     if(base_conf.empty())
         base_conf = "{}";
     rapidjson::ParseResult result = base.Parse(base_conf.data());
-    if(result)
-    {
-        for(auto iter = base.MemberBegin(); iter != base.MemberEnd(); iter++)
-            json.AddMember(iter->name, iter->value, alloc);
-    }
-    else
+    if (!result)
         writeLog(0, std::string("SIP008 base loader failed with error: ") + rapidjson::GetParseError_En(result.Code()) + " (" + std::to_string(result.Offset()) + ")", LOG_LEVEL_ERROR);
 
-    rapidjson::Value jsondata;
-    jsondata = json.Move();
-
-    output_content = "[";
+    rapidjson::Value proxies(rapidjson::kArrayType);
     for(Proxy &x : nodes)
     {
-        remark = x.Remark;
-        hostname = x.Hostname;
+        std::string &remark = x.Remark;
+        std::string &hostname = x.Hostname;
         std::string &password = x.Password;
         std::string &method = x.EncryptMethod;
         std::string &plugin = x.Plugin;
@@ -990,29 +1166,28 @@ std::string proxyToSSSub(std::string base_conf, std::vector<Proxy> &nodes, extra
         default:
             continue;
         }
-        jsondata["remarks"].SetString(rapidjson::StringRef(remark.c_str(), remark.size()));
-        jsondata["server"].SetString(rapidjson::StringRef(hostname.c_str(), hostname.size()));
-        jsondata["server_port"] = x.Port;
-        jsondata["password"].SetString(rapidjson::StringRef(password.c_str(), password.size()));
-        jsondata["method"].SetString(rapidjson::StringRef(method.c_str(), method.size()));
-        jsondata["plugin"].SetString(rapidjson::StringRef(plugin.c_str(), plugin.size()));
-        jsondata["plugin_opts"].SetString(rapidjson::StringRef(pluginopts.c_str(), pluginopts.size()));
-        output_content += SerializeObject(jsondata) + ",";
+        rapidjson::Value proxy(rapidjson::kObjectType);
+        proxy.CopyFrom(base, alloc)
+        | AddMemberOrReplace("remarks", rapidjson::Value(remark.c_str(), remark.size()), alloc)
+        | AddMemberOrReplace("server", rapidjson::Value(hostname.c_str(), hostname.size()), alloc)
+        | AddMemberOrReplace("server_port", rapidjson::Value(x.Port), alloc)
+        | AddMemberOrReplace("method", rapidjson::Value(method.c_str(), method.size()), alloc)
+        | AddMemberOrReplace("password", rapidjson::Value(password.c_str(), password.size()), alloc)
+        | AddMemberOrReplace("plugin", rapidjson::Value(plugin.c_str(), plugin.size()), alloc)
+        | AddMemberOrReplace("plugin_opts", rapidjson::Value(pluginopts.c_str(), pluginopts.size()), alloc);
+        proxies.PushBack(proxy, alloc);
     }
-    if(output_content.size() > 1)
-        output_content.erase(output_content.size() - 1);
-    output_content += "]";
-    return output_content;
+    return proxies | SerializeObject();
 }
 
 std::string proxyToQuan(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
     INIReader ini;
     ini.store_any_line = true;
-    if(!ext.nodelist && ini.Parse(base_conf) != 0)
+    if(!ext.nodelist && ini.parse(base_conf) != 0)
     {
-        writeLog(0, "Quantumult base loader failed with error: " + ini.GetLastError(), LOG_LEVEL_ERROR);
-        return std::string();
+        writeLog(0, "Quantumult base loader failed with error: " + ini.get_last_error(), LOG_LEVEL_ERROR);
+        return "";
     }
 
     proxyToQuan(nodes, ini, ruleset_content_array, extra_proxy_group, ext);
@@ -1021,38 +1196,36 @@ std::string proxyToQuan(std::vector<Proxy> &nodes, const std::string &base_conf,
     {
         string_array allnodes;
         std::string allLinks;
-        ini.GetAll("SERVER", "{NONAME}", allnodes);
+        ini.get_all("SERVER", "{NONAME}", allnodes);
         if(!allnodes.empty())
             allLinks = join(allnodes, "\n");
         return base64Encode(allLinks);
     }
-    return ini.ToString();
+    return ini.to_string();
 }
 
 void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
-    std::string type, proxyStr;
-    tribool scv;
+    std::string proxyStr;
     std::vector<Proxy> nodelist;
     string_array remarks_list;
 
-    ini.SetCurrentSection("SERVER");
-    ini.EraseSection();
+    ini.set_current_section("SERVER");
+    ini.erase_section();
     for(Proxy &x : nodes)
     {
-        std::string remark = x.Remark;
-
         if(ext.append_proxy_type)
         {
             std::string type = getProxyTypeName(x.Type);
             x.Remark = "[" + type + "] " + x.Remark;
         }
 
-        processRemark(x.Remark, remark, remarks_list);
+        processRemark(x.Remark, remarks_list);
 
         std::string &hostname = x.Hostname, &method = x.EncryptMethod, &password = x.Password, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &path = x.Path, &edge = x.Edge, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam, &plugin = x.Plugin, &pluginopts = x.PluginOption, &username = x.Username;
         std::string port = std::to_string(x.Port);
         bool &tlssecure = x.TLSSecure;
+        tribool scv;
 
         switch(x.Type)
         {
@@ -1062,7 +1235,7 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
 
             if(method == "auto")
                 method = "chacha20-ietf-poly1305";
-            proxyStr = remark + " = vmess, " + hostname + ", " + port + ", " + method + ", \"" + id + "\", group=" + x.Group;
+            proxyStr = x.Remark + " = vmess, " + hostname + ", " + port + ", " + method + ", \"" + id + "\", group=" + x.Group;
             if(tlssecure)
             {
                 proxyStr += ", over-tls=true, tls-host=" + host;
@@ -1084,12 +1257,12 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
             if(ext.nodelist)
             {
                 proxyStr = "ssr://" + urlSafeBase64Encode(hostname + ":" + port + ":" + protocol + ":" + method + ":" + obfs + ":" + urlSafeBase64Encode(password) \
-                           + "/?group=" + urlSafeBase64Encode(x.Group) + "&remarks=" + urlSafeBase64Encode(remark) \
+                           + "/?group=" + urlSafeBase64Encode(x.Group) + "&remarks=" + urlSafeBase64Encode(x.Remark) \
                            + "&obfsparam=" + urlSafeBase64Encode(obfsparam) + "&protoparam=" + urlSafeBase64Encode(protoparam));
             }
             else
             {
-                proxyStr = remark + " = shadowsocksr, " + hostname + ", " + port + ", " + method + ", \"" + password + "\", group=" + x.Group + ", protocol=" + protocol + ", obfs=" + obfs;
+                proxyStr = x.Remark + " = shadowsocksr, " + hostname + ", " + port + ", " + method + ", \"" + password + "\", group=" + x.Group + ", protocol=" + protocol + ", obfs=" + obfs;
                 if(!protoparam.empty())
                     proxyStr += ", protocol_param=" + protoparam;
                 if(!obfsparam.empty())
@@ -1104,11 +1277,11 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
                 {
                     proxyStr += "/?plugin=" + urlEncode(plugin + ";" + pluginopts);
                 }
-                proxyStr += "&group=" + urlSafeBase64Encode(x.Group) + "#" + urlEncode(remark);
+                proxyStr += "&group=" + urlSafeBase64Encode(x.Group) + "#" + urlEncode(x.Remark);
             }
             else
             {
-                proxyStr = remark + " = shadowsocks, " + hostname + ", " + port + ", " + method + ", \"" + password + "\", group=" + x.Group;
+                proxyStr = x.Remark + " = shadowsocks, " + hostname + ", " + port + ", " + method + ", \"" + password + "\", group=" + x.Group;
                 if(plugin == "obfs-local" && !pluginopts.empty())
                 {
                     proxyStr += ", " + replaceAllDistinct(pluginopts, ";", ", ");
@@ -1117,7 +1290,7 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
             break;
         case ProxyType::HTTP:
         case ProxyType::HTTPS:
-            proxyStr = remark + " = http, upstream-proxy-address=" + hostname + ", upstream-proxy-port=" + port + ", group=" + x.Group;
+            proxyStr = x.Remark + " = http, upstream-proxy-address=" + hostname + ", upstream-proxy-port=" + port + ", group=" + x.Group;
             if(!username.empty() && !password.empty())
                 proxyStr += ", upstream-proxy-auth=true, upstream-proxy-username=" + username + ", upstream-proxy-password=" + password;
             else
@@ -1136,7 +1309,7 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
                 proxyStr = "http://" + urlSafeBase64Encode(proxyStr);
             break;
         case ProxyType::SOCKS5:
-            proxyStr = remark + " = socks, upstream-proxy-address=" + hostname + ", upstream-proxy-port=" + port + ", group=" + x.Group;
+            proxyStr = x.Remark + " = socks, upstream-proxy-address=" + hostname + ", upstream-proxy-port=" + port + ", group=" + x.Group;
             if(!username.empty() && !password.empty())
                 proxyStr += ", upstream-proxy-auth=true, upstream-proxy-username=" + username + ", upstream-proxy-password=" + password;
             else
@@ -1158,23 +1331,23 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
             continue;
         }
 
-        ini.Set("{NONAME}", proxyStr);
-        remarks_list.emplace_back(std::move(remark));
+        ini.set("{NONAME}", proxyStr);
+        remarks_list.emplace_back(x.Remark);
         nodelist.emplace_back(x);
     }
 
     if(ext.nodelist)
         return;
 
-    string_array filtered_nodelist;
-    ini.SetCurrentSection("POLICY");
-    ini.EraseSection();
+    ini.set_current_section("POLICY");
+    ini.erase_section();
 
-    std::string singlegroup;
-    std::string name, proxies;
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
-        eraseElements(filtered_nodelist);
+        string_array filtered_nodelist;
+        std::string type;
+        std::string singlegroup;
+        std::string name, proxies;
 
         switch(x.Type)
         {
@@ -1204,7 +1377,7 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
                 if(!celluar.empty())
                     singlegroup += ", celluar = " + celluar;
                 singlegroup += "\n" + replaceAllDistinct(trimOf(content, ','), ",", "\n");
-                ini.Set("{NONAME}", base64Encode(singlegroup)); //insert order
+                ini.set("{NONAME}", base64Encode(singlegroup)); //insert order
             }
             continue;
         default:
@@ -1226,28 +1399,28 @@ void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetC
         if(type == "static")
             singlegroup += ", " + filtered_nodelist[0];
         singlegroup += "\n" + proxies + "\n";
-        ini.Set("{NONAME}", base64Encode(singlegroup));
+        ini.set("{NONAME}", base64Encode(singlegroup));
     }
 
     if(ext.enable_rule_generator)
-        rulesetToSurge(ini, ruleset_content_array, -2, ext.overwrite_original_rules, std::string());
+        rulesetToSurge(ini, ruleset_content_array, -2, ext.overwrite_original_rules, "");
 }
 
 std::string proxyToQuanX(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
     INIReader ini;
     ini.store_any_line = true;
-    ini.AddDirectSaveSection("general");
-    ini.AddDirectSaveSection("dns");
-    ini.AddDirectSaveSection("rewrite_remote");
-    ini.AddDirectSaveSection("rewrite_local");
-    ini.AddDirectSaveSection("task_local");
-    ini.AddDirectSaveSection("mitm");
-    ini.AddDirectSaveSection("server_remote");
-    if(!ext.nodelist && ini.Parse(base_conf) != 0)
+    ini.add_direct_save_section("general");
+    ini.add_direct_save_section("dns");
+    ini.add_direct_save_section("rewrite_remote");
+    ini.add_direct_save_section("rewrite_local");
+    ini.add_direct_save_section("task_local");
+    ini.add_direct_save_section("mitm");
+    ini.add_direct_save_section("server_remote");
+    if(!ext.nodelist && ini.parse(base_conf) != 0)
     {
-        writeLog(0, "QuantumultX base loader failed with error: " + ini.GetLastError(), LOG_LEVEL_ERROR);
-        return std::string();
+        writeLog(0, "QuantumultX base loader failed with error: " + ini.get_last_error(), LOG_LEVEL_ERROR);
+        return "";
     }
 
     proxyToQuanX(nodes, ini, ruleset_content_array, extra_proxy_group, ext);
@@ -1256,34 +1429,32 @@ std::string proxyToQuanX(std::vector<Proxy> &nodes, const std::string &base_conf
     {
         string_array allnodes;
         std::string allLinks;
-        ini.GetAll("server_local", "{NONAME}", allnodes);
+        ini.get_all("server_local", "{NONAME}", allnodes);
         if(!allnodes.empty())
             allLinks = join(allnodes, "\n");
         return allLinks;
     }
-    return ini.ToString();
+    return ini.to_string();
 }
 
 void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
-    std::string type;
-    std::string remark, hostname, port, method;
-    std::string password, plugin, pluginopts;
-    std::string id, transproto, host, path;
-    std::string protocol, protoparam, obfs, obfsparam;
     std::string proxyStr;
     tribool udp, tfo, scv, tls13;
     std::vector<Proxy> nodelist;
     string_array remarks_list;
 
-    ini.SetCurrentSection("server_local");
-    ini.EraseSection();
+    ini.set_current_section("server_local");
+    ini.erase_section();
     for(Proxy &x : nodes)
     {
         if(ext.append_proxy_type)
+        {
+            std::string type = getProxyTypeName(x.Type);
             x.Remark = "[" + type + "] " + x.Remark;
+        }
 
-        processRemark(x.Remark, remark, remarks_list);
+        processRemark(x.Remark, remarks_list);
 
         std::string &hostname = x.Hostname, &method = x.EncryptMethod, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &path = x.Path, &password = x.Password, &plugin = x.Plugin, &pluginopts = x.PluginOption, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam, &username = x.Username;
         std::string port = std::to_string(x.Port);
@@ -1303,7 +1474,9 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
         case ProxyType::VMess:
             if(method == "auto")
                 method = "chacha20-ietf-poly1305";
-            proxyStr = "vmess = " + hostname + ":" + port + ", method=" + method + ", password=" + id + ", aead=" + (x.AlterId == 0 ? "true" : "false");
+            proxyStr = "vmess = " + hostname + ":" + port + ", method=" + method + ", password=" + id;
+            if (x.AlterId != 0)
+                proxyStr += ", aead=false";
             if(tlssecure && !tls13.is_undef())
                 proxyStr += ", tls13=" + std::string(tls13 ? "true" : "false");
             if(transproto == "ws")
@@ -1333,7 +1506,7 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
                         plugin = getUrlArg(pluginopts, "mode") == "websocket" ? "ws" : "";
                         host = getUrlArg(pluginopts, "host");
                         path = getUrlArg(pluginopts, "path");
-                        tlssecure = pluginopts.find("tls") != pluginopts.npos;
+                        tlssecure = pluginopts.find("tls") != std::string::npos;
                         if(tlssecure && plugin == "ws")
                         {
                             plugin += 's';
@@ -1412,10 +1585,10 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
             proxyStr += ", udp-relay=" + udp.get_str();
         if(tlssecure && !scv.is_undef() && (x.Type != ProxyType::Shadowsocks && x.Type != ProxyType::ShadowsocksR))
             proxyStr += ", tls-verification=" + scv.reverse().get_str();
-        proxyStr += ", tag=" + remark;
+        proxyStr += ", tag=" + x.Remark;
 
-        ini.Set("{NONAME}", proxyStr);
-        remarks_list.emplace_back(std::move(remark));
+        ini.set("{NONAME}", proxyStr);
+        remarks_list.emplace_back(x.Remark);
         nodelist.emplace_back(x);
     }
 
@@ -1423,12 +1596,13 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
         return;
 
     string_multimap original_groups;
-    ini.SetCurrentSection("policy");
-    ini.GetItems(original_groups);
-    ini.EraseSection();
+    ini.set_current_section("policy");
+    ini.get_items(original_groups);
+    ini.erase_section();
 
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
+        std::string type;
         string_array filtered_nodelist;
 
         switch(x.Type)
@@ -1447,8 +1621,8 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
             break;
         case ProxyGroupType::SSID:
             type = "ssid";
-            for(auto iter = x.Proxies.begin(); iter != x.Proxies.end(); iter++)
-                filtered_nodelist.emplace_back(replaceAllDistinct(*iter, "=", ":"));
+            for(const auto & proxy : x.Proxies)
+                filtered_nodelist.emplace_back(replaceAllDistinct(proxy, "=", ":"));
             break;
         default:
             continue;
@@ -1469,8 +1643,8 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
         auto iter = std::find_if(original_groups.begin(), original_groups.end(), [&](const string_multimap::value_type &n)
         {
             std::string groupdata = n.second;
-            std::string::size_type cpos = groupdata.find(",");
-            if(cpos != groupdata.npos)
+            std::string::size_type cpos = groupdata.find(',');
+            if(cpos != std::string::npos)
                 return trim(groupdata.substr(0, cpos)) == x.Name;
             else
                 return false;
@@ -1488,7 +1662,13 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
         std::string proxies = join(filtered_nodelist, ", ");
 
         std::string singlegroup = type + "=" + x.Name + ", " + proxies;
-        ini.Set("{NONAME}", singlegroup);
+        if(x.Type != ProxyGroupType::Select && x.Type != ProxyGroupType::SSID)
+        {
+            singlegroup += ", check-interval=" + std::to_string(x.Interval);
+            if(x.Tolerance > 0)
+                singlegroup += ", tolerance=" + std::to_string(x.Tolerance);
+        }
+        ini.set("{NONAME}", singlegroup);
     }
 
     if(ext.enable_rule_generator)
@@ -1499,7 +1679,7 @@ std::string proxyToSSD(std::vector<Proxy> &nodes, std::string &group, std::strin
 {
     rapidjson::StringBuffer sb;
     rapidjson::Writer<rapidjson::StringBuffer> writer(sb);
-    size_t index = 0;
+    int index = 0;
 
     if(group.empty())
         group = "SSD";
@@ -1598,29 +1778,29 @@ std::string proxyToMellow(std::vector<Proxy> &nodes, const std::string &base_con
 {
     INIReader ini;
     ini.store_any_line = true;
-    if(ini.Parse(base_conf) != 0)
+    if(ini.parse(base_conf) != 0)
     {
-        writeLog(0, "Mellow base loader failed with error: " + ini.GetLastError(), LOG_LEVEL_ERROR);
-        return std::string();
+        writeLog(0, "Mellow base loader failed with error: " + ini.get_last_error(), LOG_LEVEL_ERROR);
+        return "";
     }
 
     proxyToMellow(nodes, ini, ruleset_content_array, extra_proxy_group, ext);
 
-    return ini.ToString();
+    return ini.to_string();
 }
 
 void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
     std::string proxy;
-    std::string type, remark, hostname, port, username, password, method;
+    std::string username, password, method;
     std::string plugin, pluginopts;
     std::string id, aid, transproto, faketype, host, path, quicsecure, quicsecret, tlssecure;
     std::string url;
     tribool tfo, scv;
     std::vector<Proxy> nodelist;
-    string_array vArray, remarks_list, filtered_nodelist;
+    string_array vArray, remarks_list;
 
-    ini.SetCurrentSection("Endpoint");
+    ini.set_current_section("Endpoint");
 
     for(Proxy &x : nodes)
     {
@@ -1630,10 +1810,9 @@ void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Rulese
             x.Remark = "[" + type + "] " + x.Remark;
         }
 
-        processRemark(x.Remark, remark, remarks_list);
+        processRemark(x.Remark, remarks_list);
 
-        std::string &hostname = x.Hostname;
-        port = std::to_string(x.Port);
+        std::string &hostname = x.Hostname, port = std::to_string(x.Port);
 
         tfo = ext.tfo;
         scv = ext.skip_cert_verify;
@@ -1645,10 +1824,10 @@ void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Rulese
         case ProxyType::Shadowsocks:
             if(!x.Plugin.empty())
                 continue;
-            proxy = remark + ", ss, ss://" + urlSafeBase64Encode(method + ":" + password) + "@" + hostname + ":" + port;
+            proxy = x.Remark + ", ss, ss://" + urlSafeBase64Encode(method + ":" + password) + "@" + hostname + ":" + port;
             break;
         case ProxyType::VMess:
-            proxy = remark + ", vmess1, vmess1://" + id + "@" + hostname + ":" + port;
+            proxy = x.Remark + ", vmess1, vmess1://" + id + "@" + hostname + ":" + port;
             if(!path.empty())
                 proxy += path;
             proxy += "?network=" + transproto;
@@ -1666,7 +1845,6 @@ void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Rulese
                     proxy += "&quic.security=" + quicsecure + "&quic.key=" + quicsecret;
                 break;
             case "kcp"_hash:
-                break;
             case "tcp"_hash:
                 break;
             }
@@ -1682,25 +1860,25 @@ void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Rulese
                 proxy += "&sockopt.tcpfastopen=" + tfo.get_str();
             break;
         case ProxyType::SOCKS5:
-            proxy = remark + ", builtin, socks, address=" + hostname + ", port=" + port + ", user=" + username + ", pass=" + password;
+            proxy = x.Remark + ", builtin, socks, address=" + hostname + ", port=" + port + ", user=" + username + ", pass=" + password;
             break;
         case ProxyType::HTTP:
-            proxy = remark + ", builtin, http, address=" + hostname + ", port=" + port + ", user=" + username + ", pass=" + password;
+            proxy = x.Remark + ", builtin, http, address=" + hostname + ", port=" + port + ", user=" + username + ", pass=" + password;
             break;
         default:
             continue;
         }
 
-        ini.Set("{NONAME}", proxy);
-        remarks_list.emplace_back(std::move(remark));
+        ini.set("{NONAME}", proxy);
+        remarks_list.emplace_back(x.Remark);
         nodelist.emplace_back(x);
     }
 
-    ini.SetCurrentSection("EndpointGroup");
+    ini.set_current_section("EndpointGroup");
 
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
-        eraseElements(filtered_nodelist);
+        string_array filtered_nodelist;
         url.clear();
         proxy.clear();
 
@@ -1744,34 +1922,31 @@ void proxyToMellow(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Rulese
         proxy += join(filtered_nodelist, ":");
         proxy += ", latency, interval=300, timeout=6"; //use hard-coded values for now
 
-        ini.Set("{NONAME}", proxy); //insert order
+        ini.set("{NONAME}", proxy); //insert order
     }
 
     if(ext.enable_rule_generator)
-        rulesetToSurge(ini, ruleset_content_array, 0, ext.overwrite_original_rules, std::string());
+        rulesetToSurge(ini, ruleset_content_array, 0, ext.overwrite_original_rules, "");
 }
 
 std::string proxyToLoon(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
 {
-    rapidjson::Document json;
     INIReader ini;
-    std::string proxy;
     std::string output_nodelist;
-    tribool scv;
     std::vector<Proxy> nodelist;
 
-    string_array vArray, remarks_list, filtered_nodelist;
+    string_array remarks_list;
 
     ini.store_any_line = true;
-    if(ini.Parse(base_conf) != INIREADER_EXCEPTION_NONE && !ext.nodelist)
+    ini.add_direct_save_section("Plugin");
+    if(ini.parse(base_conf) != INIREADER_EXCEPTION_NONE && !ext.nodelist)
     {
-        writeLog(0, "Loon base loader failed with error: " + ini.GetLastError(), LOG_LEVEL_ERROR);
-        return std::string();
+        writeLog(0, "Loon base loader failed with error: " + ini.get_last_error(), LOG_LEVEL_ERROR);
+        return "";
     }
 
-
-    ini.SetCurrentSection("Proxy");
-    ini.EraseSection();
+    ini.set_current_section("Proxy");
+    ini.erase_section();
 
     for(Proxy &x : nodes)
     {
@@ -1780,8 +1955,7 @@ std::string proxyToLoon(std::vector<Proxy> &nodes, const std::string &base_conf,
             std::string type = getProxyTypeName(x.Type);
             x.Remark = "[" + type + "] " + x.Remark;
         }
-        std::string remark = x.Remark;
-        processRemark(x.Remark, remark, remarks_list);
+        processRemark(x.Remark, remarks_list);
 
         std::string &hostname = x.Hostname, &username = x.Username, &password = x.Password, &method = x.EncryptMethod, &plugin = x.Plugin, &pluginopts = x.PluginOption, &id = x.UserId, &transproto = x.TransferProtocol, &host = x.Host, &path = x.Path, &protocol = x.Protocol, &protoparam = x.ProtocolParam, &obfs = x.OBFS, &obfsparam = x.OBFSParam;
         std::string port = std::to_string(x.Port), aid = std::to_string(x.AlterId);
@@ -1790,7 +1964,7 @@ std::string proxyToLoon(std::vector<Proxy> &nodes, const std::string &base_conf,
         tribool scv = ext.skip_cert_verify;
         scv.define(x.AllowInsecure);
 
-        proxy.clear();
+        std::string proxy;
 
         switch(x.Type)
         {
@@ -1808,86 +1982,120 @@ std::string proxyToLoon(std::vector<Proxy> &nodes, const std::string &base_conf,
             if(method == "auto")
                 method = "chacha20-ietf-poly1305";
 
-            proxy = "vmess," + hostname + "," + port + "," + method + ",\"" + id + "\",over-tls:" + (tlssecure ? "true" : "false");
+            proxy = "vmess," + hostname + "," + port + "," + method + ",\"" + id + "\",over-tls=" + (tlssecure ? "true" : "false");
             if(tlssecure)
-                proxy += ",tls-name:" + host;
+                proxy += ",tls-name=" + host;
             switch(hash_(transproto))
             {
             case "tcp"_hash:
-                proxy += ",transport:tcp";
+                proxy += ",transport=tcp";
                 break;
             case "ws"_hash:
-                proxy += ",transport:ws,path:" + path + ",host:" + host;
+                proxy += ",transport=ws,path=" + path + ",host=" + host;
                 break;
             default:
                 continue;
             }
             if(!scv.is_undef())
-                proxy += ",skip-cert-verify:" + std::string(scv.get() ? "1" : "0");
+                proxy += ",skip-cert-verify=" + std::string(scv.get() ? "true" : "false");
             break;
         case ProxyType::ShadowsocksR:
-            proxy = "ShadowsocksR," + hostname + "," + port + "," + method + ",\"" + password + "\"," + protocol + ",{" + protoparam + "}," + obfs + ",{" + obfsparam + "}";
+            proxy = "ShadowsocksR," + hostname + "," + port + "," + method + ",\"" + password + "\",protocol=" + protocol + ",protocol-param=" + protoparam + ",obfs=" + obfs + ",obfs-param=" + obfsparam;
             break;
-        /*
-        case ProxyType::SOCKS5:
-            proxy = "socks5, " + hostname + ", " + port + ", " + username + ", " + password;
-            if(ext.skip_cert_verify)
-                proxy += ", skip-cert-verify:1";
-            break;
-        */
         case ProxyType::HTTP:
-            proxy = "http," + hostname + "," + port + "," + username + "," + password;
+            proxy = "http," + hostname + "," + port + "," + username + ",\"" + password + "\"";
+            break;
+        case ProxyType::HTTPS:
+            proxy = "https," + hostname + "," + port + "," + username + ",\"" + password + "\"";
+            if(!host.empty())
+                proxy += ",tls-name=" + host;
+            if(!scv.is_undef())
+                proxy += ",skip-cert-verify=" + std::string(scv.get() ? "true" : "false");
             break;
         case ProxyType::Trojan:
-            proxy = "trojan," + hostname + "," + port + "," + password;
+            proxy = "trojan," + hostname + "," + port + ",\"" + password + "\"";
             if(!host.empty())
-                proxy += ",tls-name:" + host;
+                proxy += ",tls-name=" + host;
             if(!scv.is_undef())
-                proxy += ",skip-cert-verify:" + std::string(scv.get() ? "1" : "0");
+                proxy += ",skip-cert-verify=" + std::string(scv.get() ? "true" : "false");
+            break;
+        case ProxyType::SOCKS5:
+            proxy = "socks5," + hostname + "," + port;
+            if (!username.empty() && !password.empty())
+                proxy += "," + username + ",\"" + password + "\"";
+            proxy += ",over-tls=" + std::string(tlssecure ? "true" : "false");
+            if (tlssecure)
+            {
+                if(!host.empty())
+                    proxy += ",tls-name=" + host;
+                if(!scv.is_undef())
+                    proxy += ",skip-cert-verify=" + std::string(scv.get() ? "true" : "false");
+            }
+            break;
+        case ProxyType::WireGuard:
+            proxy = "wireguard, interface-ip=" + x.SelfIP;
+            if(!x.SelfIPv6.empty())
+                proxy += ", interface-ipv6=" + x.SelfIPv6;
+            proxy += ", private-key=" + x.PrivateKey;
+            for(const auto &y : x.DnsServers)
+            {
+                if(isIPv4(y))
+                    proxy += ", dns=" + y;
+                else if(isIPv6(y))
+                    proxy += ", dnsv6=" + y;
+            }
+            if(x.Mtu > 0)
+                proxy += ", mtu=" + std::to_string(x.Mtu);
+            if(x.KeepAlive > 0)
+                proxy += ", keepalive=" + std::to_string(x.KeepAlive);
+            proxy += ", peers=[{" + generatePeer(x, true) + "}]";
             break;
         default:
             continue;
         }
 
-        /*
         if(ext.tfo)
-            proxy += ", tfo=true";
+            proxy += ",fast-open=true";
         if(ext.udp)
-            proxy += ", udp-relay=true";
-        */
+            proxy += ",udp=true";
+
 
         if(ext.nodelist)
-            output_nodelist += remark + " = " + proxy + "\n";
+            output_nodelist += x.Remark + " = " + proxy + "\n";
         else
         {
-            ini.Set("{NONAME}", remark + " = " + proxy);
+            ini.set("{NONAME}", x.Remark + " = " + proxy);
             nodelist.emplace_back(x);
-            remarks_list.emplace_back(std::move(remark));
+            remarks_list.emplace_back(x.Remark);
         }
     }
 
     if(ext.nodelist)
         return output_nodelist;
 
-    ini.SetCurrentSection("Proxy Group");
-    ini.EraseSection();
+    string_multimap original_groups;
+    ini.set_current_section("Proxy Group");
+    ini.get_items(original_groups);
+    ini.erase_section();
+
     for(const ProxyGroupConfig &x : extra_proxy_group)
     {
-        eraseElements(filtered_nodelist);
-        proxy.clear();
+        string_array filtered_nodelist;
+        std::string group, group_extra;
 
         switch(x.Type)
         {
         case ProxyGroupType::Select:
+        case ProxyGroupType::LoadBalance:
         case ProxyGroupType::URLTest:
         case ProxyGroupType::Fallback:
             break;
         case ProxyGroupType::SSID:
             if(x.Proxies.size() < 2)
                 continue;
-            proxy = x.TypeStr() + ",default=" + x.Proxies[0] + ",";
-            proxy += join(x.Proxies.begin() + 1, x.Proxies.end(), ",");
-            ini.Set("{NONAME}", x.Name + " = " + proxy); //insert order
+            group = x.TypeStr() + ",default=" + x.Proxies[0] + ",";
+            group += join(x.Proxies.begin() + 1, x.Proxies.end(), ",");
+                ini.set("{NONAME}", x.Name + " = " + group); //insert order
             continue;
         default:
             continue;
@@ -1899,20 +2107,471 @@ std::string proxyToLoon(std::vector<Proxy> &nodes, const std::string &base_conf,
         if(filtered_nodelist.empty())
             filtered_nodelist.emplace_back("DIRECT");
 
-        proxy = x.TypeStr() + ",";
+        auto iter = std::find_if(original_groups.begin(), original_groups.end(), [&](const string_multimap::value_type &n)
+        {
+            return trim(n.first) == x.Name;
+        });
+
+        if(iter != original_groups.end())
+        {
+            string_array vArray = split(iter->second, ",");
+            if(vArray.size() > 1)
+            {
+                if(trim(vArray[vArray.size() - 1]).find("img-url") == 0)
+                    filtered_nodelist.emplace_back(trim(vArray[vArray.size() - 1]));
+            }
+        }
+
+        group = x.TypeStr() + ",";
         /*
         for(std::string &y : filtered_nodelist)
-            proxy += "," + y;
+            group += "," + y;
         */
-        proxy += join(filtered_nodelist, ",");
-        if(x.Type == ProxyGroupType::URLTest || x.Type == ProxyGroupType::Fallback)
-            proxy += ",url=" + x.Url + ",interval=" + std::to_string(x.Interval);
+        group += join(filtered_nodelist, ",");
+        if(x.Type != ProxyGroupType::Select)
+        {
+            group += ",url=" + x.Url + ",interval=" + std::to_string(x.Interval);
+            if(x.Type == ProxyGroupType::LoadBalance)
+            {
+                group += ",algorithm=" + std::string(x.Strategy == BalanceStrategy::RoundRobin ? "round-robin" : "pcc");
+                if(x.Timeout > 0)
+                    group += ",max-timeout=" + std::to_string(x.Timeout);
+            }
+            if(x.Type == ProxyGroupType::URLTest)
+            {
+                if(x.Tolerance > 0)
+                    group += ",tolerance=" + std::to_string(x.Tolerance);
+            }
+            if(x.Type == ProxyGroupType::Fallback)
+                group += ",max-timeout=" + std::to_string(x.Timeout);
+        }
 
-        ini.Set("{NONAME}", x.Name + " = " + proxy); //insert order
+        ini.set("{NONAME}", x.Name + " = " + group); //insert order
     }
 
     if(ext.enable_rule_generator)
         rulesetToSurge(ini, ruleset_content_array, -4, ext.overwrite_original_rules, ext.managed_config_prefix);
 
-    return ini.ToString();
+    return ini.to_string();
+}
+
+static std::string formatSingBoxInterval(Integer interval)
+{
+    std::string result;
+    if(interval >= 3600)
+    {
+        result += std::to_string(interval / 3600) + "h";
+        interval %= 3600;
+    }
+    if(interval >= 60)
+    {
+        result += std::to_string(interval / 60) + "m";
+        interval %= 60;
+    }
+    if(interval > 0)
+        result += std::to_string(interval) + "s";
+    return result;
+}
+
+static rapidjson::Value buildSingBoxTransport(const Proxy& proxy, rapidjson::MemoryPoolAllocator<>& allocator)
+{
+    rapidjson::Value transport(rapidjson::kObjectType);
+    switch (hash_(proxy.TransferProtocol))
+    {
+        case "http"_hash:
+        {
+            if (!proxy.Host.empty())
+                transport.AddMember("host", rapidjson::StringRef(proxy.Host.c_str()), allocator);
+            [[fallthrough]];
+        }
+        case "ws"_hash:
+        {
+            transport.AddMember("type", rapidjson::StringRef(proxy.TransferProtocol.c_str()), allocator);
+            if (proxy.Path.empty())
+                transport.AddMember("path", "/", allocator);
+            else
+                transport.AddMember("path", rapidjson::StringRef(proxy.Path.c_str()), allocator);
+
+            rapidjson::Value headers(rapidjson::kObjectType);
+            if (!proxy.Host.empty())
+                headers.AddMember("Host", rapidjson::StringRef(proxy.Host.c_str()), allocator);
+            if (!proxy.Edge.empty())
+                headers.AddMember("Edge", rapidjson::StringRef(proxy.Edge.c_str()), allocator);
+            transport.AddMember("headers", headers, allocator);
+            break;
+        }
+        case "grpc"_hash:
+        {
+            transport.AddMember("type", "grpc", allocator);
+            if (!proxy.Path.empty())
+                transport.AddMember("service_name", rapidjson::StringRef(proxy.Path.c_str()), allocator);
+            break;
+        }
+        default:
+            break;
+    }
+    return transport;
+}
+
+static void addSingBoxCommonMembers(rapidjson::Value &proxy, const Proxy &x, const rapidjson::GenericStringRef<rapidjson::Value::Ch> &type, rapidjson::MemoryPoolAllocator<> &allocator)
+{
+    proxy.AddMember("type", type, allocator);
+    proxy.AddMember("tag", rapidjson::StringRef(x.Remark.c_str()), allocator);
+    proxy.AddMember("server", rapidjson::StringRef(x.Hostname.c_str()), allocator);
+    proxy.AddMember("server_port", x.Port, allocator);
+}
+
+static rapidjson::Value stringArrayToJsonArray(const std::string &array, const std::string &delimiter, rapidjson::MemoryPoolAllocator<> &allocator)
+{
+    rapidjson::Value result(rapidjson::kArrayType);
+    string_array vArray = split(array, delimiter);
+    for (const auto &x : vArray)
+        result.PushBack(rapidjson::Value(trim(x).c_str(), allocator), allocator);
+    return result;
+}
+
+void proxyToSingBox(std::vector<Proxy> &nodes, rapidjson::Document &json, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext) {
+    using namespace rapidjson_ext;
+    rapidjson::Document::AllocatorType &allocator = json.GetAllocator();
+    rapidjson::Value outbounds(rapidjson::kArrayType), route(rapidjson::kArrayType);
+    std::vector<Proxy> nodelist;
+    string_array remarks_list;
+
+    if (!ext.nodelist)
+    {
+        auto direct = buildObject(allocator, "type", "direct", "tag", "DIRECT");
+        outbounds.PushBack(direct, allocator);
+        auto reject = buildObject(allocator, "type", "block", "tag", "REJECT");
+        outbounds.PushBack(reject, allocator);
+        auto dns = buildObject(allocator, "type", "dns", "tag", "dns-out");
+        outbounds.PushBack(dns, allocator);
+    }
+
+    for (Proxy &x : nodes)
+    {
+        std::string type = getProxyTypeName(x.Type);
+        if (ext.append_proxy_type)
+            x.Remark = "[" + type + "] " + x.Remark;
+
+        processRemark(x.Remark, remarks_list, false);
+
+        tribool udp = ext.udp, tfo = ext.tfo, scv = ext.skip_cert_verify;
+        udp.define(x.UDP);
+        tfo.define(x.TCPFastOpen);
+        scv.define(x.AllowInsecure);
+        rapidjson::Value proxy(rapidjson::kObjectType);
+        switch (x.Type)
+        {
+            case ProxyType::Shadowsocks:
+            {
+                addSingBoxCommonMembers(proxy, x, "shadowsocks", allocator);
+                proxy.AddMember("method", rapidjson::StringRef(x.EncryptMethod.c_str()), allocator);
+                proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+                if(!x.Plugin.empty() && !x.PluginOption.empty())
+                {
+                    if (x.Plugin == "simple-obfs")
+                        x.Plugin = "obfs-local";
+                    proxy.AddMember("plugin", rapidjson::StringRef(x.Plugin.c_str()), allocator);
+                    proxy.AddMember("plugin_opts", rapidjson::StringRef(x.PluginOption.c_str()), allocator);
+                }
+                break;
+            }
+            case ProxyType::ShadowsocksR:
+            {
+                addSingBoxCommonMembers(proxy, x, "shadowsocksr", allocator);
+                proxy.AddMember("method", rapidjson::StringRef(x.EncryptMethod.c_str()), allocator);
+                proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+                proxy.AddMember("protocol", rapidjson::StringRef(x.Protocol.c_str()), allocator);
+                proxy.AddMember("protocol_param", rapidjson::StringRef(x.ProtocolParam.c_str()), allocator);
+                proxy.AddMember("obfs", rapidjson::StringRef(x.OBFS.c_str()), allocator);
+                proxy.AddMember("obfs_param", rapidjson::StringRef(x.OBFSParam.c_str()), allocator);
+                break;
+            }
+            case ProxyType::VMess:
+            {
+                addSingBoxCommonMembers(proxy, x, "vmess", allocator);
+                proxy.AddMember("uuid", rapidjson::StringRef(x.UserId.c_str()), allocator);
+                proxy.AddMember("alter_id", x.AlterId, allocator);
+                proxy.AddMember("security", rapidjson::StringRef(x.EncryptMethod.c_str()), allocator);
+
+                auto transport = buildSingBoxTransport(x, allocator);
+                if (!transport.ObjectEmpty())
+                    proxy.AddMember("transport", transport, allocator);
+                break;
+            }
+            case ProxyType::Trojan:
+            {
+                addSingBoxCommonMembers(proxy, x, "trojan", allocator);
+                proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+
+                auto transport = buildSingBoxTransport(x, allocator);
+                if (!transport.ObjectEmpty())
+                    proxy.AddMember("transport", transport, allocator);
+                break;
+            }
+            case ProxyType::WireGuard:
+            {
+                proxy.AddMember("type", "wireguard", allocator);
+                proxy.AddMember("tag", rapidjson::StringRef(x.Remark.c_str()), allocator);
+                rapidjson::Value addresses(rapidjson::kArrayType);
+                addresses.PushBack(rapidjson::StringRef(x.SelfIP.c_str()), allocator);
+                if (!x.SelfIPv6.empty())
+                    addresses.PushBack(rapidjson::StringRef(x.SelfIPv6.c_str()), allocator);
+                proxy.AddMember("local_address", addresses, allocator);
+                proxy.AddMember("private_key", rapidjson::StringRef(x.PrivateKey.c_str()), allocator);
+
+                rapidjson::Value peer(rapidjson::kObjectType);
+                peer.AddMember("server", rapidjson::StringRef(x.Hostname.c_str()), allocator);
+                peer.AddMember("server_port", x.Port, allocator);
+                peer.AddMember("public_key", rapidjson::StringRef(x.PublicKey.c_str()), allocator);
+                if (!x.PreSharedKey.empty())
+                    peer.AddMember("pre_shared_key", rapidjson::StringRef(x.PreSharedKey.c_str()), allocator);
+
+                if (!x.AllowedIPs.empty())
+                {
+                    auto allowed_ips = stringArrayToJsonArray(x.AllowedIPs, ",", allocator);
+                    peer.AddMember("allowed_ips", allowed_ips, allocator);
+                }
+
+                if (!x.ClientId.empty())
+                {
+                    auto reserved = stringArrayToJsonArray(x.ClientId, ",", allocator);
+                    peer.AddMember("reserved", reserved, allocator);
+                }
+
+                rapidjson::Value peers(rapidjson::kArrayType);
+                peers.PushBack(peer, allocator);
+                proxy.AddMember("peers", peers, allocator);
+                proxy.AddMember("mtu", x.Mtu, allocator);
+                break;
+            }
+            case ProxyType::Hysteria:
+            {
+                addSingBoxCommonMembers(proxy, x, "hysteria", allocator);
+                if (!x.Up.empty())
+                    proxy.AddMember("up_mbps", x.UpSpeed, allocator);
+                if (!x.Down.empty())
+                    proxy.AddMember("down_mbps", x.DownSpeed, allocator);
+                if (!x.OBFS.empty())
+                {
+                    proxy.AddMember("obfs", rapidjson::StringRef(x.OBFS.c_str()), allocator);
+                }
+                
+                if (!x.AuthStr.empty())
+                {
+                    proxy.AddMember("auth_str", rapidjson::StringRef(x.AuthStr.c_str()), allocator);
+                    rapidjson::Value auth_str;
+                    auth_str.SetString(base64Encode(x.AuthStr).c_str(), allocator);
+                    proxy.AddMember("auth", auth_str, allocator);
+                }
+                if (x.RecvWindowConn)
+                    proxy.AddMember("recv_window_conn", x.RecvWindowConn, allocator);
+                if (x.RecvWindow)
+                    proxy.AddMember("recv_window", x.RecvWindow, allocator);
+                if (!x.DisableMtuDiscovery.is_undef())
+                    proxy.AddMember("disable_mtu_discovery", x.DisableMtuDiscovery.get(), allocator);
+                
+                rapidjson::Value tls(rapidjson::kObjectType);
+                tls.AddMember("enabled", true, allocator);
+                if (!scv.is_undef())
+                    tls.AddMember("insecure", scv.get(), allocator);
+                if (!x.Alpn.empty())
+                {
+                    rapidjson::Value alpn(rapidjson::kArrayType);
+                    alpn.PushBack(rapidjson::StringRef(x.Alpn[0].c_str()), allocator);
+                    tls.AddMember("alpn", alpn, allocator);
+                }
+                if (!x.Ca.empty())
+                {
+                    rapidjson::Value ca_str;
+                    ca_str.SetString(x.Ca.c_str(), allocator);
+                    tls.AddMember("certificate", ca_str, allocator);
+                }
+                if (!x.CaStr.empty())
+                    tls.AddMember("certificate", rapidjson::StringRef(x.CaStr.c_str()), allocator);
+                proxy.AddMember("tls", tls, allocator);
+                break;
+            }
+            case ProxyType::Hysteria2:
+            {
+                addSingBoxCommonMembers(proxy, x, "hysteria2", allocator);
+                if (!x.Up.empty())
+                    proxy.AddMember("up_mbps", x.UpSpeed, allocator);
+                if (!x.Down.empty())
+                    proxy.AddMember("down_mbps", x.DownSpeed, allocator);
+                if (!x.OBFS.empty())
+                {
+                    rapidjson::Value obfs(rapidjson::kObjectType);
+                    obfs.AddMember("type", rapidjson::StringRef(x.OBFS.c_str()), allocator);
+                    if (!x.OBFSParam.empty())
+                        obfs.AddMember("password", rapidjson::StringRef(x.OBFSParam.c_str()), allocator);
+                    proxy.AddMember("obfs", obfs, allocator);
+                }
+                if (!x.Password.empty())
+                    proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+                
+                rapidjson::Value tls(rapidjson::kObjectType);
+                tls.AddMember("enabled", true, allocator);
+                if (!scv.is_undef())
+                    tls.AddMember("insecure", scv.get(), allocator);
+                if (!x.Alpn.empty())
+                {
+                    rapidjson::Value alpn(rapidjson::kArrayType);
+                    alpn.PushBack(rapidjson::StringRef(x.Alpn[0].c_str()), allocator);
+                    tls.AddMember("alpn", alpn, allocator);
+                }
+                if (!x.Ca.empty())
+                {
+                    rapidjson::Value ca_str(rapidjson::kStringType);
+                    ca_str.SetString(x.Ca.c_str(), allocator);
+                    tls.AddMember("certificate", ca_str, allocator);
+                }
+                if (!x.CaStr.empty())
+                    tls.AddMember("certificate", rapidjson::StringRef(x.CaStr.c_str()), allocator);
+                proxy.AddMember("tls", tls, allocator);
+                break;
+            }
+            case ProxyType::HTTP:
+            case ProxyType::HTTPS:
+            {
+                addSingBoxCommonMembers(proxy, x, "http", allocator);
+                proxy.AddMember("username", rapidjson::StringRef(x.Username.c_str()), allocator);
+                proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+                break;
+            }
+            case ProxyType::SOCKS5:
+            {
+                addSingBoxCommonMembers(proxy, x, "socks", allocator);
+                proxy.AddMember("version", "5", allocator);
+                proxy.AddMember("username", rapidjson::StringRef(x.Username.c_str()), allocator);
+                proxy.AddMember("password", rapidjson::StringRef(x.Password.c_str()), allocator);
+                break;
+            }
+            default:
+                continue;
+        }
+        if (x.TLSSecure)
+        {
+            rapidjson::Value tls(rapidjson::kObjectType);
+            tls.AddMember("enabled", true, allocator);
+            if (!x.ServerName.empty())
+                tls.AddMember("server_name", rapidjson::StringRef(x.ServerName.c_str()), allocator);
+            else if (!x.Host.empty())
+                tls.AddMember("server_name", rapidjson::StringRef(x.Host.c_str()), allocator);
+            tls.AddMember("insecure", buildBooleanValue(scv), allocator);
+            proxy.AddMember("tls", tls, allocator);
+        }
+        if (!udp.is_undef() && !udp)
+        {
+            proxy.AddMember("network", "tcp", allocator);
+        }
+        if (!tfo.is_undef())
+        {
+            proxy.AddMember("tcp_fast_open", buildBooleanValue(tfo), allocator);
+        }
+        nodelist.push_back(x);
+        remarks_list.emplace_back(x.Remark);
+        outbounds.PushBack(proxy, allocator);
+    }
+
+    if (ext.nodelist)
+    {
+        json | AddMemberOrReplace("outbounds", outbounds, allocator);
+        return;
+    }
+
+    for (const ProxyGroupConfig &x: extra_proxy_group)
+    {
+        string_array filtered_nodelist;
+        std::string type;
+        switch (x.Type)
+        {
+            case ProxyGroupType::Select:
+            {
+                type = "selector";
+                break;
+            }
+            case ProxyGroupType::URLTest:
+            case ProxyGroupType::Fallback:
+            case ProxyGroupType::LoadBalance:
+            {
+                type = "urltest";
+                break;
+            }
+            default:
+                continue;
+        }
+        for (const auto &y : x.Proxies)
+            groupGenerate(y, nodelist, filtered_nodelist, true, ext);
+
+        if (filtered_nodelist.empty())
+            filtered_nodelist.emplace_back("DIRECT");
+
+        rapidjson::Value group(rapidjson::kObjectType);
+
+        group.AddMember("type", rapidjson::Value(type.c_str(), allocator), allocator);
+        group.AddMember("tag", rapidjson::Value(x.Name.c_str(), allocator), allocator);
+
+        rapidjson::Value group_outbounds(rapidjson::kArrayType);
+        for (const std::string& y: filtered_nodelist)
+        {
+            group_outbounds.PushBack(rapidjson::Value(y.c_str(), allocator), allocator);
+        }
+        group.AddMember("outbounds", group_outbounds, allocator);
+
+        if (x.Type == ProxyGroupType::URLTest)
+        {
+            group.AddMember("url", rapidjson::Value(x.Url.c_str(), allocator), allocator);
+            group.AddMember("interval", rapidjson::Value(formatSingBoxInterval(x.Interval).c_str(), allocator), allocator);
+            if (x.Tolerance > 0)
+                group.AddMember("tolerance", x.Tolerance, allocator);
+        }
+        outbounds.PushBack(group, allocator);
+    }
+
+    if (global.singBoxAddClashModes)
+    {
+        auto global_group = rapidjson::Value(rapidjson::kObjectType);
+        global_group.AddMember("type", "selector", allocator);
+        global_group.AddMember("tag", "GLOBAL", allocator);
+        global_group.AddMember("outbounds", rapidjson::Value(rapidjson::kArrayType), allocator);
+        global_group["outbounds"].PushBack("DIRECT", allocator);
+        for (auto &x: remarks_list)
+        {
+            global_group["outbounds"].PushBack(rapidjson::Value(x.c_str(), allocator), allocator);
+        }
+        outbounds.PushBack(global_group, allocator);
+    }
+
+    json | AddMemberOrReplace("outbounds", outbounds, allocator);
+}
+
+std::string proxyToSingBox(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext)
+{
+    using namespace rapidjson_ext;
+    rapidjson::Document json;
+
+    if (!ext.nodelist)
+    {
+        json.Parse(base_conf.data());
+        if (json.HasParseError())
+        {
+            writeLog(0, "sing-box base loader failed with error: " +
+                        std::string(rapidjson::GetParseError_En(json.GetParseError())), LOG_LEVEL_ERROR);
+            return "";
+        }
+    }
+    else
+    {
+        json.SetObject();
+    }
+
+    proxyToSingBox(nodes, json, ruleset_content_array, extra_proxy_group, ext);
+
+    if(ext.nodelist || !ext.enable_rule_generator)
+        return json | SerializeObject();
+
+    rulesetToSingBox(json, ruleset_content_array, ext.overwrite_original_rules);
+
+    return json | SerializeObject();
 }
