@@ -7,12 +7,12 @@
 #include <quickjspp.hpp>
 #endif // NO_JS_RUNTIME
 
-#include "../../config/proxygroup.h"
-#include "../../config/regmatch.h"
-#include "../../parser/config/proxy.h"
-#include "../../utils/ini_reader/ini_reader.h"
-#include "../../utils/string.h"
-#include "../../utils/yamlcpp_extra.h"
+#include "config/proxygroup.h"
+#include "config/regmatch.h"
+#include "parser/config/proxy.h"
+#include "utils/ini_reader/ini_reader.h"
+#include "utils/string.h"
+#include "utils/yamlcpp_extra.h"
 #include "ruleconvert.h"
 
 struct extra_settings
@@ -37,11 +37,12 @@ struct extra_settings
     tribool skip_cert_verify = tribool();
     tribool tls13 = tribool();
     bool clash_classical_ruleset = false;
-    std::string sort_script = "";
+    std::string sort_script;
     std::string clash_proxies_style = "flow";
+    std::string clash_proxy_groups_style = "flow";
     bool authorized = false;
 
-    extra_settings() {};
+    extra_settings() = default;
     extra_settings(const extra_settings&) = delete;
     extra_settings(extra_settings&&) = delete;
 
@@ -57,9 +58,6 @@ struct extra_settings
 #endif // NO_JS_RUNTIME
 };
 
-void rulesetToClash(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, bool overwrite_original_rules, bool new_field_name);
-void rulesetToSurge(INIReader &base_rule, std::vector<RulesetContent> &ruleset_content_array, int surge_ver, bool overwrite_original_rules, std::string remote_path_prefix);
-
 std::string proxyToClash(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, bool clashR, extra_settings &ext);
 void proxyToClash(std::vector<Proxy> &nodes, YAML::Node &yamlnode, const ProxyGroupConfigs &extra_proxy_group, bool clashR, extra_settings &ext);
 std::string proxyToSurge(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, int surge_ver, extra_settings &ext);
@@ -73,5 +71,6 @@ void proxyToQuanX(std::vector<Proxy> &nodes, INIReader &ini, std::vector<Ruleset
 std::string proxyToQuan(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext);
 void proxyToQuan(std::vector<Proxy> &nodes, INIReader &ini, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext);
 std::string proxyToSSD(std::vector<Proxy> &nodes, std::string &group, std::string &userinfo, extra_settings &ext);
+std::string proxyToSingBox(std::vector<Proxy> &nodes, const std::string &base_conf, std::vector<RulesetContent> &ruleset_content_array, const ProxyGroupConfigs &extra_proxy_group, extra_settings &ext);
 
 #endif // SUBEXPORT_H_INCLUDED
