@@ -4,8 +4,8 @@
 #include <string>
 #include <map>
 
-#include "../../generator/config/subexport.h"
-#include "../../utils/string.h"
+#include "generator/config/subexport.h"
+#include "utils/string.h"
 
 struct template_args
 {
@@ -16,6 +16,6 @@ struct template_args
 };
 
 int render_template(const std::string &content, const template_args &vars, std::string &output, const std::string &include_scope = "templates");
-int renderClashScript(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, std::string remote_path_prefix, bool script, bool overwrite_original_rules, bool clash_classic_ruleset);
+int renderClashScript(YAML::Node &base_rule, std::vector<RulesetContent> &ruleset_content_array, const std::string &remote_path_prefix, bool script, bool overwrite_original_rules, bool clash_classic_ruleset);
 
 #endif // TEMPLATES_H_INCLUDED
