@@ -3,30 +3,26 @@
 #include <mutex>
 #include <numeric>
 
-#include <inja.hpp>
 #include <yaml-cpp/yaml.h>
 
-#include "../config/binding.h"
-#include "../generator/config/nodemanip.h"
-#include "../generator/config/ruleconvert.h"
-#include "../generator/config/subexport.h"
-#include "../generator/template/templates.h"
-#include "../script/cron.h"
-#include "../script/script_quickjs.h"
-#include "../server/webserver.h"
-#include "../utils/base64/base64.h"
-#include "../utils/file_extra.h"
-#include "../utils/ini_reader/ini_reader.h"
-#include "../utils/logger.h"
-#include "../utils/network.h"
-#include "../utils/regexp.h"
-#include "../utils/stl_extra.h"
-#include "../utils/string.h"
-#include "../utils/string_hash.h"
-#include "../utils/system.h"
-#include "../utils/system.h"
-#include "../utils/urlencode.h"
-#include "../utils/yamlcpp_extra.h"
+#include "config/binding.h"
+#include "generator/config/nodemanip.h"
+#include "generator/config/ruleconvert.h"
+#include "generator/config/subexport.h"
+#include "generator/template/templates.h"
+#include "script/script_quickjs.h"
+#include "server/webserver.h"
+#include "utils/base64/base64.h"
+#include "utils/file_extra.h"
+#include "utils/ini_reader/ini_reader.h"
+#include "utils/logger.h"
+#include "utils/network.h"
+#include "utils/regexp.h"
+#include "utils/stl_extra.h"
+#include "utils/string.h"
+#include "utils/string_hash.h"
+#include "utils/system.h"
+#include "utils/urlencode.h"
 #include "interfaces.h"
 #include "multithread.h"
 #include "settings.h"
@@ -36,8 +32,6 @@
 extern WebServer webServer;
 
 string_array gRegexBlacklist = {"(.*)*"};
-
-void refreshRulesets(RulesetConfigs &ruleset_list, std::vector<RulesetContent> &ruleset_content_array);
 
 std::string parseProxy(const std::string &source)
 {
@@ -67,6 +61,7 @@ const std::vector<UAProfile> UAMatchList = {
     {"ClashForAndroid","","","clash",false},
     {"ClashforWindows","\\/([0-9.]+)","0.11","clash",true},
     {"ClashforWindows","","","clash",false},
+    {"clash-verge","","","clash",true},
     {"ClashX Pro","","","clash",true},
     {"ClashX","\\/([0-9.]+)","0.13","clash",true},
     {"Clash","","","clash",true},
@@ -89,30 +84,28 @@ const std::vector<UAProfile> UAMatchList = {
     {"V2RayX","","","v2ray"}
 };
 
-bool verGreaterEqual(const std::string &src_ver, const std::string &target_ver)
+bool verGreaterEqual(const std::string& src_ver, const std::string& target_ver)
 {
-    string_size src_pos_beg = 0, src_pos_end, target_pos_beg = 0, target_pos_end;
-    while(true)
-    {
-        src_pos_end = src_ver.find('.', src_pos_beg);
-        if(src_pos_end == src_ver.npos)
-            src_pos_end = src_ver.size();
-        int part_src = std::stoi(src_ver.substr(src_pos_beg, src_pos_end - src_pos_beg));
-        target_pos_end = target_ver.find('.', target_pos_beg);
-        if(target_pos_end == target_ver.npos)
-            target_pos_end = target_ver.size();
-        int part_target = std::stoi(target_ver.substr(target_pos_beg, target_pos_end - target_pos_beg));
-        if(part_src > part_target)
-            break;
-        else if(part_src < part_target)
-            return false;
-        else if(src_pos_end >= src_ver.size() - 1 || target_pos_end >= target_ver.size() - 1)
-            break;
-        src_pos_beg = src_pos_end + 1;
-        target_pos_beg = target_pos_end + 1;
+    std::istringstream src_stream(src_ver), target_stream(target_ver);
+    int src_part, target_part;
+    char dot;
+    while (src_stream >> src_part) {
+        if (target_stream >> target_part) {
+            if (src_part < target_part) {
+                return false;
+            } else if (src_part > target_part) {
+                return true;
+            }
+            // Skip the dot separator in both streams
+            src_stream >> dot;
+            target_stream >> dot;
+        } else {
+            // If we run out of target parts, the source version is greater only if it has more parts
+            return true;
+        }
     }
-    return true;
-
+    // If we get here, the common parts are equal, so check if target_ver has more parts
+    return !bool(target_stream >> target_part);
 }
 
 void matchUserAgent(const std::string &user_agent, std::string &target, tribool &clash_new_name, int &surge_ver)
@@ -138,25 +131,18 @@ void matchUserAgent(const std::string &user_agent, std::string &target, tribool 
             return;
         }
     }
-    return;
-}
-
-std::string getConvertedRuleset(RESPONSE_CALLBACK_ARGS)
-{
-    std::string url = urlDecode(getUrlArg(request.argument, "url")), type = getUrlArg(request.argument, "type");
-    return convertRuleset(fetchFile(url, parseProxy(global.proxyRuleset), global.cacheRuleset), to_int(type));
 }
 
 std::string getRuleset(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto &argument = request.argument;
     int *status_code = &response.status_code;
     /// type: 1 for Surge, 2 for Quantumult X, 3 for Clash domain rule-provider, 4 for Clash ipcidr rule-provider, 5 for Surge DOMAIN-SET, 6 for Clash classical ruleset
     std::string url = urlSafeBase64Decode(getUrlArg(argument, "url")), type = getUrlArg(argument, "type"), group = urlSafeBase64Decode(getUrlArg(argument, "group"));
     std::string output_content, dummy;
     int type_int = to_int(type, 0);
 
-    if(!url.size() || !type.size() || (type_int == 2 && !group.size()) || (type_int < 1 || type_int > 6))
+    if(url.empty() || type.empty() || (type_int == 2 && group.empty()) || (type_int < 1 || type_int > 6))
     {
         *status_code = 400;
         return "Invalid request!";
@@ -175,7 +161,7 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
         output_content += convertRuleset(content, x.rule_type);
     }
 
-    if(!output_content.size())
+    if(output_content.empty())
     {
         *status_code = 400;
         return "Invalid request!";
@@ -192,11 +178,11 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
     {
         posb = 0;
         pose = strLine.find(',');
-        if(pose == strLine.npos)
+        if(pose == std::string::npos)
             return 1;
         posb = pose + 1;
         pose = strLine.find(',', posb);
-        if(pose == strLine.npos)
+        if(pose == std::string::npos)
         {
             pose = strLine.size();
             if(strLine[pose - 1] == '\r')
@@ -223,11 +209,11 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
         switch(type_int)
         {
         case 2:
-            if(!std::any_of(QuanXRuleTypes.begin(), QuanXRuleTypes.end(), [&strLine](std::string type){return startsWith(strLine, type);}))
+            if(!std::any_of(QuanXRuleTypes.begin(), QuanXRuleTypes.end(), [&strLine](const std::string& type){return startsWith(strLine, type);}))
                 continue;
             break;
         case 1:
-            if(!std::any_of(SurgeRuleTypes.begin(), SurgeRuleTypes.end(), [&strLine](std::string type){return startsWith(strLine, type);}))
+            if(!std::any_of(SurgeRuleTypes.begin(), SurgeRuleTypes.end(), [&strLine](const std::string& type){return startsWith(strLine, type);}))
                 continue;
             break;
         case 3:
@@ -238,7 +224,7 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
             output_content += "  - '";
             if(strLine[posb - 2] == 'X')
                 output_content += "+.";
-            output_content += strLine.substr(posb, pose);
+            output_content += trim(strLine.substr(posb, pose));
             output_content += "'\n";
             continue;
         case 4:
@@ -247,7 +233,7 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
             if(filterLine())
                 continue;
             output_content += "  - '";
-            output_content += strLine.substr(posb, pose);
+            output_content += trim(strLine.substr(posb, pose));
             output_content += "'\n";
             continue;
         case 5:
@@ -255,13 +241,17 @@ std::string getRuleset(RESPONSE_CALLBACK_ARGS)
                 continue;
             if(filterLine())
                 continue;
-            output_content += strLine.substr(posb, pose);
+            if(strLine[posb - 2] == 'X')
+                output_content += '.';
+            output_content += trim(strLine.substr(posb, pose));
             output_content += '\n';
             continue;
         case 6:
-            if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](std::string type){return startsWith(strLine, type);}))
+            if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](const std::string& type){return startsWith(strLine, type);}))
                 continue;
             output_content += "  - ";
+        default:
+            break;
         }
 
         lineSize = strLine.size();
@@ -311,12 +301,12 @@ void checkExternalBase(const std::string &path, std::string &dest)
 
 std::string subconverter(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto &argument = request.argument;
     int *status_code = &response.status_code;
 
     std::string argTarget = getUrlArg(argument, "target"), argSurgeVer = getUrlArg(argument, "ver");
     tribool argClashNewField = getUrlArg(argument, "new_name");
-    int intSurgeVer = argSurgeVer.size() ? to_int(argSurgeVer, 3) : 3;
+    int intSurgeVer = !argSurgeVer.empty() ? to_int(argSurgeVer, 3) : 3;
     if(argTarget == "auto")
         matchUserAgent(request.headers["User-Agent"], argTarget, argClashNewField, intSurgeVer);
 
@@ -327,23 +317,23 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     case "ss"_hash: case "ssd"_hash: case "ssr"_hash: case "sssub"_hash: case "v2ray"_hash: case "trojan"_hash: case "mixed"_hash:
         lSimpleSubscription = true;
         break;
-    case "clash"_hash: case "clashr"_hash: case "surge"_hash: case "quan"_hash: case "quanx"_hash: case "loon"_hash: case "surfboard"_hash: case "mellow"_hash:
+    case "clash"_hash: case "clashr"_hash: case "surge"_hash: case "quan"_hash: case "quanx"_hash: case "loon"_hash: case "surfboard"_hash: case "mellow"_hash: case "singbox"_hash:
         break;
     default:
         *status_code = 400;
         return "Invalid target!";
     }
     //check if we need to read configuration
-    if((!global.APIMode || global.CFWChildProcess) && !global.generatorMode)
+    if(global.reloadConfOnRequest && (!global.APIMode || global.CFWChildProcess) && !global.generatorMode)
         readConf();
 
     /// string values
-    std::string argUrl = urlDecode(getUrlArg(argument, "url"));
-    std::string argGroupName = urlDecode(getUrlArg(argument, "group")), argUploadPath = getUrlArg(argument, "upload_path");
-    std::string argIncludeRemark = urlDecode(getUrlArg(argument, "include")), argExcludeRemark = urlDecode(getUrlArg(argument, "exclude"));
-    std::string argCustomGroups = urlSafeBase64Decode(getUrlArg(argument, "groups")), argCustomRulesets = urlSafeBase64Decode(getUrlArg(argument, "ruleset")), argExternalConfig = urlDecode(getUrlArg(argument, "config"));
-    std::string argDeviceID = getUrlArg(argument, "dev_id"), argFilename = urlDecode(getUrlArg(argument, "filename")), argUpdateInterval = getUrlArg(argument, "interval"), argUpdateStrict = getUrlArg(argument, "strict");
-    std::string argRenames = urlDecode(getUrlArg(argument, "rename")), argFilterScript = urlDecode(getUrlArg(argument, "filter_script"));
+    std::string argUrl = getUrlArg(argument, "url");
+    std::string argGroupName = getUrlArg(argument, "group"), argUploadPath = getUrlArg(argument, "upload_path");
+    std::string argIncludeRemark = getUrlArg(argument, "include"), argExcludeRemark = getUrlArg(argument, "exclude");
+    std::string argCustomGroups = urlSafeBase64Decode(getUrlArg(argument, "groups")), argCustomRulesets = urlSafeBase64Decode(getUrlArg(argument, "ruleset")), argExternalConfig = getUrlArg(argument, "config");
+    std::string argDeviceID = getUrlArg(argument, "dev_id"), argFilename = getUrlArg(argument, "filename"), argUpdateInterval = getUrlArg(argument, "interval"), argUpdateStrict = getUrlArg(argument, "strict");
+    std::string argRenames = getUrlArg(argument, "rename"), argFilterScript = getUrlArg(argument, "filter_script");
 
     /// switches with default value
     tribool argUpload = getUrlArg(argument, "upload"), argEmoji = getUrlArg(argument, "emoji"), argAddEmoji = getUrlArg(argument, "add_emoji"), argRemoveEmoji = getUrlArg(argument, "remove_emoji");
@@ -360,8 +350,8 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     std::vector<RulesetContent> lRulesetContent;
     extra_settings ext;
     std::string subInfo, dummy;
-    int interval = argUpdateInterval.size() ? to_int(argUpdateInterval, global.updateInterval) : global.updateInterval;
-    bool authorized = !global.APIMode || getUrlArg(argument, "token") == global.accessToken, strict = argUpdateStrict.size() ? argUpdateStrict == "true" : global.updateStrict;
+    int interval = !argUpdateInterval.empty() ? to_int(argUpdateInterval, global.updateInterval) : global.updateInterval;
+    bool authorized = !global.APIMode || getUrlArg(argument, "token") == global.accessToken, strict = !argUpdateStrict.empty() ? argUpdateStrict == "true" : global.updateStrict;
 
     if(std::find(gRegexBlacklist.cbegin(), gRegexBlacklist.cend(), argIncludeRemark) != gRegexBlacklist.cend() || std::find(gRegexBlacklist.cbegin(), gRegexBlacklist.cend(), argExcludeRemark) != gRegexBlacklist.cend())
         return "Invalid request!";
@@ -369,31 +359,39 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     /// for external configuration
     std::string lClashBase = global.clashBase, lSurgeBase = global.surgeBase, lMellowBase = global.mellowBase, lSurfboardBase = global.surfboardBase;
     std::string lQuanBase = global.quanBase, lQuanXBase = global.quanXBase, lLoonBase = global.loonBase, lSSSubBase = global.SSSubBase;
+    std::string lSingBoxBase = global.singBoxBase;
 
     /// validate urls
     argEnableInsert.define(global.enableInsert);
-    if(!argUrl.size() && (!global.APIMode || authorized))
+    if(argUrl.empty() && (!global.APIMode || authorized))
         argUrl = global.defaultUrls;
-    if((!argUrl.size() && !(global.insertUrls.size() && argEnableInsert)) || !argTarget.size())
+    if((argUrl.empty() && !(!global.insertUrls.empty() && argEnableInsert)) || argTarget.empty())
     {
         *status_code = 400;
         return "Invalid request!";
     }
 
     /// load request arguments as template variables
-    string_array req_args = split(argument, "&");
+//    string_array req_args = split(argument, "&");
+//    string_map req_arg_map;
+//    for(std::string &x : req_args)
+//    {
+//        string_size pos = x.find("=");
+//        if(pos == x.npos)
+//        {
+//            req_arg_map[x] = "";
+//            continue;
+//        }
+//        if(x.substr(0, pos) == "token")
+//            continue;
+//        req_arg_map[x.substr(0, pos)] = x.substr(pos + 1);
+//    }
     string_map req_arg_map;
-    for(std::string &x : req_args)
+    for (auto &x : argument)
     {
-        string_size pos = x.find("=");
-        if(pos == x.npos)
-        {
-            req_arg_map[x] = "";
+        if(x.first == "token")
             continue;
-        }
-        if(x.substr(0, pos) == "token")
-            continue;
-        req_arg_map[x.substr(0, pos)] = x.substr(pos + 1);
+        req_arg_map[x.first] = x.second;
     }
     req_arg_map["target"] = argTarget;
     req_arg_map["ver"] = std::to_string(intSurgeVer);
@@ -413,6 +411,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
         argExpandRulesets.define(true);
 
     ext.clash_proxies_style = global.clashProxiesStyle;
+    ext.clash_proxy_groups_style = global.clashProxyGroupsStyle;
 
     /// read preference from argument, assign global var if not in argument
     ext.tfo.define(argTFO).define(global.TFOFlag);
@@ -421,7 +420,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     ext.tls13.define(argTLS13).define(global.TLS13Flag);
 
     ext.sort_flag = argSort.get(global.enableSort);
-    argUseSortScript.define(global.sortScript.size() != 0);
+    argUseSortScript.define(!global.sortScript.empty());
     if(ext.sort_flag && argUseSortScript)
         ext.sort_script = global.sortScript;
     ext.filter_deprecated = argFilterDeprecated.get(global.filterDeprecated);
@@ -435,7 +434,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
 
     ext.nodelist = argGenNodeList;
     ext.surge_ssr_path = global.surgeSSRPath;
-    ext.quanx_dev_id = argDeviceID.size() ? argDeviceID : global.quanXDevID;
+    ext.quanx_dev_id = !argDeviceID.empty() ? argDeviceID : global.quanXDevID;
     ext.enable_rule_generator = global.enableRuleGen;
     ext.overwrite_original_rules = global.overwriteOriginalRules;
     if(!argExpandRulesets)
@@ -444,7 +443,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     /// load external configuration
     if(argExternalConfig.empty())
         argExternalConfig = global.defaultExtConfig;
-    if(argExternalConfig.size())
+    if(!argExternalConfig.empty())
     {
         //std::cerr<<"External configuration file provided. Loading...\n";
         writeLog(0, "External configuration file provided. Loading...", LOG_LEVEL_INFO);
@@ -464,22 +463,23 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
                     checkExternalBase(extconf.quan_rule_base, lQuanBase);
                     checkExternalBase(extconf.quanx_rule_base, lQuanXBase);
                     checkExternalBase(extconf.loon_rule_base, lLoonBase);
+                    checkExternalBase(extconf.singbox_rule_base, lSingBoxBase);
 
-                    if(extconf.surge_ruleset.size())
+                    if(!extconf.surge_ruleset.empty())
                         lCustomRulesets = extconf.surge_ruleset;
-                    if(extconf.custom_proxy_group.size())
+                    if(!extconf.custom_proxy_group.empty())
                         lCustomProxyGroups = extconf.custom_proxy_group;
                     ext.enable_rule_generator = extconf.enable_rule_generator;
                     ext.overwrite_original_rules = extconf.overwrite_original_rules;
                 }
             }
-            if(extconf.rename.size())
+            if(!extconf.rename.empty())
                 ext.rename_array = extconf.rename;
-            if(extconf.emoji.size())
+            if(!extconf.emoji.empty())
                 ext.emoji_array = extconf.emoji;
-            if(extconf.include.size())
+            if(!extconf.include.empty())
                 lIncludeRemarks = extconf.include;
-            if(extconf.exclude.size())
+            if(!extconf.exclude.empty())
                 lExcludeRemarks = extconf.exclude;
             argAddEmoji.define(extconf.add_emoji);
             argRemoveEmoji.define(extconf.remove_old_emoji);
@@ -490,14 +490,14 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
         if(!lSimpleSubscription)
         {
             /// loading custom groups
-            if(argCustomGroups.size() && !ext.nodelist)
+            if(!argCustomGroups.empty() && !ext.nodelist)
             {
                 string_array vArray = split(argCustomGroups, "@");
                 lCustomProxyGroups = INIBinding::from<ProxyGroupConfig>::from_ini(vArray);
             }
 
             /// loading custom rulesets
-            if(argCustomRulesets.size() && !ext.nodelist)
+            if(!argCustomRulesets.empty() && !ext.nodelist)
             {
                 string_array vArray = split(argCustomRulesets, "@");
                 lCustomRulesets = INIBinding::from<RulesetConfig>::from_ini(vArray);
@@ -525,15 +525,15 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     ext.remove_emoji = argRemoveEmoji.get(global.removeEmoji);
     if(ext.add_emoji && ext.emoji_array.empty())
         ext.emoji_array = safe_get_emojis();
-    if(argRenames.size())
+    if(!argRenames.empty())
         ext.rename_array = INIBinding::from<RegexMatchConfig>::from_ini(split(argRenames, "`"), "@");
     else if(ext.rename_array.empty())
         ext.rename_array = safe_get_renames();
 
     /// check custom include/exclude settings
-    if(argIncludeRemark.size() && regValid(argIncludeRemark))
+    if(!argIncludeRemark.empty() && regValid(argIncludeRemark))
         lIncludeRemarks = string_array{argIncludeRemark};
-    if(argExcludeRemark.size() && regValid(argExcludeRemark))
+    if(!argExcludeRemark.empty() && regValid(argExcludeRemark))
         lExcludeRemarks = string_array{argExcludeRemark};
 
     /// initialize script runtime
@@ -565,7 +565,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     parse_set.js_runtime = ext.js_runtime;
     parse_set.js_context = ext.js_context;
 
-    if(global.insertUrls.size() && argEnableInsert)
+    if(!global.insertUrls.empty() && argEnableInsert)
     {
         groupID = -1;
         urls = split(global.insertUrls, "|");
@@ -608,12 +608,12 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
         groupID++;
     }
     //exit if found nothing
-    if(!nodes.size() && !insert_nodes.size())
+    if(nodes.empty() && insert_nodes.empty())
     {
         *status_code = 400;
         return "No nodes were found!";
     }
-    if(subInfo.size() && argAppendUserinfo.get(global.appendUserinfo))
+    if(!subInfo.empty() && argAppendUserinfo.get(global.appendUserinfo))
         response.headers.emplace("Subscription-UserInfo", subInfo);
 
     if(request.method == "HEAD")
@@ -633,7 +633,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     std::string filterScript = global.filterScript;
     if(authorized && !argFilterScript.empty())
         filterScript = argFilterScript;
-    if(filterScript.size())
+    if(!filterScript.empty())
     {
         if(startsWith(filterScript, "path:"))
             filterScript = fileGet(filterScript.substr(5), false);
@@ -676,7 +676,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
     }
 
     //check custom group name
-    if(argGroupName.size())
+    if(!argGroupName.empty())
         for(Proxy &x : nodes)
             x.Group = argGroupName;
 
@@ -699,9 +699,9 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
 
     ProxyGroupConfigs dummy_group;
     std::vector<RulesetContent> dummy_ruleset;
-    std::string managed_url = base64Decode(urlDecode(getUrlArg(argument, "profile_data")));
+    std::string managed_url = base64Decode(getUrlArg(argument, "profile_data"));
     if(managed_url.empty())
-        managed_url = global.managedConfigPrefix + "/sub?" + argument;
+        managed_url = global.managedConfigPrefix + "/sub?" + joinArguments(argument);
 
     //std::cerr<<"Generate target: ";
     proxy = parseProxy(global.proxyConfig);
@@ -753,7 +753,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
             if(argUpload)
                 uploadGist("surge" + argSurgeVer, argUploadPath, output_content, true);
 
-            if(global.writeManagedConfig && global.managedConfigPrefix.size())
+            if(global.writeManagedConfig && !global.managedConfigPrefix.empty())
                 output_content = "#!MANAGED-CONFIG " + managed_url + (interval ? " interval=" + std::to_string(interval) : "") \
                  + " strict=" + std::string(strict ? "true" : "false") + "\n\n" + output_content;
         }
@@ -770,7 +770,7 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
         if(argUpload)
             uploadGist("surfboard", argUploadPath, output_content, true);
 
-        if(global.writeManagedConfig && global.managedConfigPrefix.size())
+        if(global.writeManagedConfig && !global.managedConfigPrefix.empty())
             output_content = "#!MANAGED-CONFIG " + managed_url + (interval ? " interval=" + std::to_string(interval) : "") \
                  + " strict=" + std::string(strict ? "true" : "false") + "\n\n" + output_content;
         break;
@@ -883,24 +883,40 @@ std::string subconverter(RESPONSE_CALLBACK_ARGS)
         if(argUpload)
             uploadGist("ssd", argUploadPath, output_content, false);
         break;
+    case "singbox"_hash:
+        writeLog(0, "Generate target: sing-box", LOG_LEVEL_INFO);
+        if(!ext.nodelist)
+        {
+            if(render_template(fetchFile(lSingBoxBase, proxy, global.cacheConfig), tpl_args, base_content, global.templatePath) != 0)
+            {
+                *status_code = 400;
+                return base_content;
+            }
+        }
+
+        output_content = proxyToSingBox(nodes, base_content, lRulesetContent, lCustomProxyGroups, ext);
+
+        if(argUpload)
+            uploadGist("singbox", argUploadPath, output_content, false);
+        break;
     default:
         writeLog(0, "Generate target: Unspecified", LOG_LEVEL_INFO);
         *status_code = 500;
         return "Unrecognized target";
     }
     writeLog(0, "Generate completed.", LOG_LEVEL_INFO);
-    if(argFilename.size())
+    if(!argFilename.empty())
         response.headers.emplace("Content-Disposition", "attachment; filename=\"" + argFilename + "\"; filename*=utf-8''" + urlEncode(argFilename));
     return output_content;
 }
 
 std::string simpleToClashR(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto argument = joinArguments(request.argument);
     int *status_code = &response.status_code;
 
     std::string url = argument.size() <= 8 ? "" : argument.substr(8);
-    if(!url.size() || argument.substr(0, 8) != "sublink=")
+    if(url.empty() || argument.substr(0, 8) != "sublink=")
     {
         *status_code = 400;
         return "Invalid request!";
@@ -910,13 +926,14 @@ std::string simpleToClashR(RESPONSE_CALLBACK_ARGS)
         *status_code = 400;
         return "Please insert your subscription link instead of clicking the default link.";
     }
-    request.argument = "target=clashr&url=" + urlEncode(url);
+    request.argument.emplace("target", "clashr");
+    request.argument.emplace("url", url);
     return subconverter(request, response);
 }
 
 std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto argument = joinArguments(request.argument);
     int *status_code = &response.status_code;
 
     INIReader ini;
@@ -927,9 +944,9 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
 
     ini.store_any_line = true;
 
-    if(!url.size())
+    if(url.empty())
         url = global.defaultUrls;
-    if(!url.size() || argument.substr(0, 5) != "link=")
+    if(url.empty() || argument.substr(0, 5) != "link=")
     {
         *status_code = 400;
         return "Invalid request!";
@@ -958,15 +975,15 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
 
     base_content = fetchFile(url, proxy, global.cacheConfig);
 
-    if(ini.Parse(base_content) != INIREADER_EXCEPTION_NONE)
+    if(ini.parse(base_content) != INIREADER_EXCEPTION_NONE)
     {
-        std::string errmsg = "Parsing Surge config failed! Reason: " + ini.GetLastError();
+        std::string errmsg = "Parsing Surge config failed! Reason: " + ini.get_last_error();
         //std::cerr<<errmsg<<"\n";
         writeLog(0, errmsg, LOG_LEVEL_ERROR);
         *status_code = 400;
         return errmsg;
     }
-    if(!ini.SectionExist("Proxy") || !ini.SectionExist("Proxy Group") || !ini.SectionExist("Rule"))
+    if(!ini.section_exist("Proxy") || !ini.section_exist("Proxy Group") || !ini.section_exist("Rule"))
     {
         std::string errmsg = "Incomplete surge config! Missing critical sections!";
         //std::cerr<<errmsg<<"\n";
@@ -977,7 +994,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
 
     //scan groups first, get potential policy-path
     string_multimap section;
-    ini.GetItems("Proxy Group", section);
+    ini.get_items("Proxy Group", section);
     std::string name, type, content;
     string_array links;
     links.emplace_back(url);
@@ -988,7 +1005,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
         name = x.first;
         content = x.second;
         dummy_str_array = split(content, ",");
-        if(!dummy_str_array.size())
+        if(dummy_str_array.empty())
             continue;
         type = dummy_str_array[0];
         if(!(type == "select" || type == "url-test" || type == "fallback" || type == "load-balance")) //remove unsupported types
@@ -998,11 +1015,11 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
         for(unsigned int i = 1; i < dummy_str_array.size(); i++)
         {
             if(startsWith(dummy_str_array[i], "url"))
-                singlegroup["url"] = trim(dummy_str_array[i].substr(dummy_str_array[i].find("=") + 1));
+                singlegroup["url"] = trim(dummy_str_array[i].substr(dummy_str_array[i].find('=') + 1));
             else if(startsWith(dummy_str_array[i], "interval"))
-                singlegroup["interval"] = trim(dummy_str_array[i].substr(dummy_str_array[i].find("=") + 1));
+                singlegroup["interval"] = trim(dummy_str_array[i].substr(dummy_str_array[i].find('=') + 1));
             else if(startsWith(dummy_str_array[i], "policy-path"))
-                links.emplace_back(trim(dummy_str_array[i].substr(dummy_str_array[i].find("=") + 1)));
+                links.emplace_back(trim(dummy_str_array[i].substr(dummy_str_array[i].find('=') + 1)));
             else
                 singlegroup["proxies"].push_back(trim(dummy_str_array[i]));
         }
@@ -1038,7 +1055,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
     }
 
     //exit if found nothing
-    if(!nodes.size())
+    if(nodes.empty())
     {
         *status_code = 400;
         return "No nodes were found!";
@@ -1053,19 +1070,20 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
     ext.skip_cert_verify = global.skipCertVerify;
     ext.tls13 = global.TLS13Flag;
     ext.clash_proxies_style = global.clashProxiesStyle;
+    ext.clash_proxy_groups_style = global.clashProxyGroupsStyle;
 
     ProxyGroupConfigs dummy_groups;
     proxyToClash(nodes, clash, dummy_groups, false, ext);
 
     section.clear();
-    ini.GetItems("Proxy", section);
+    ini.get_items("Proxy", section);
     for(auto &x : section)
     {
         singlegroup.reset();
         name = x.first;
         content = x.second;
         dummy_str_array = split(content, ",");
-        if(!dummy_str_array.size())
+        if(dummy_str_array.empty())
             continue;
         content = trim(dummy_str_array[0]);
         switch(hash_(content))
@@ -1088,7 +1106,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
     }
 
     eraseElements(dummy_str_array);
-    ini.GetAll("Rule", "{NONAME}", dummy_str_array);
+    ini.get_all("Rule", "{NONAME}", dummy_str_array);
     YAML::Node rule;
     string_array strArray;
     std::string strLine;
@@ -1102,7 +1120,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
             if(strArray.size() != 3)
                 continue;
             content = webGet(strArray[1], proxy, global.cacheRuleset);
-            if(!content.size())
+            if(content.empty())
                 continue;
 
             ss << content;
@@ -1115,7 +1133,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
                     strLine.erase(--lineSize);
                 if(!lineSize || strLine[0] == ';' || strLine[0] == '#' || (lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')) //empty lines and comments are ignored
                     continue;
-                else if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](std::string type){return startsWith(strLine, type);})) //remove unsupported types
+                else if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](const std::string& type){return startsWith(strLine, type);})) //remove unsupported types
                     continue;
                 strLine += strArray[2];
                 if(count_least(strLine, ',', 3))
@@ -1125,7 +1143,7 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
             ss.clear();
             continue;
         }
-        else if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](std::string type){return startsWith(strLine, type);}))
+        else if(!std::any_of(ClashRuleTypes.begin(), ClashRuleTypes.end(), [&strLine](const std::string& type){return startsWith(strLine, type);}))
             continue;
         rule.push_back(x);
     }
@@ -1138,23 +1156,24 @@ std::string surgeConfToClash(RESPONSE_CALLBACK_ARGS)
 
 std::string getProfile(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto &argument = request.argument;
     int *status_code = &response.status_code;
 
-    std::string name = urlDecode(getUrlArg(argument, "name")), token = urlDecode(getUrlArg(argument, "token"));
+    std::string name = getUrlArg(argument, "name"), token = getUrlArg(argument, "token");
     string_array profiles = split(name, "|");
-    name = profiles[0];
-    if(token.empty() || name.empty())
+    if(token.empty() || profiles.empty())
     {
         *status_code = 403;
         return "Forbidden";
     }
     std::string profile_content;
+    name = profiles[0];
     /*if(vfs::vfs_exist(name))
     {
         profile_content = vfs::vfs_get(name);
     }
-    else */if(fileExist(name))
+    else */
+    if(fileExist(name))
     {
         profile_content = fileGet(name, true);
     }
@@ -1166,18 +1185,18 @@ std::string getProfile(RESPONSE_CALLBACK_ARGS)
     //std::cerr<<"Trying to load profile '" + name + "'.\n";
     writeLog(0, "Trying to load profile '" + name + "'.", LOG_LEVEL_INFO);
     INIReader ini;
-    if(ini.Parse(profile_content) != INIREADER_EXCEPTION_NONE && !ini.SectionExist("Profile"))
+    if(ini.parse(profile_content) != INIREADER_EXCEPTION_NONE && !ini.section_exist("Profile"))
     {
-        //std::cerr<<"Load profile failed! Reason: "<<ini.GetLastError()<<"\n";
-        writeLog(0, "Load profile failed! Reason: " + ini.GetLastError(), LOG_LEVEL_ERROR);
+        //std::cerr<<"Load profile failed! Reason: "<<ini.get_last_error()<<"\n";
+        writeLog(0, "Load profile failed! Reason: " + ini.get_last_error(), LOG_LEVEL_ERROR);
         *status_code = 500;
         return "Broken profile!";
     }
     //std::cerr<<"Trying to parse profile '" + name + "'.\n";
     writeLog(0, "Trying to parse profile '" + name + "'.", LOG_LEVEL_INFO);
     string_multimap contents;
-    ini.GetItems("Profile", contents);
-    if(!contents.size())
+    ini.get_items("Profile", contents);
+    if(contents.empty())
     {
         //std::cerr<<"Load profile failed! Reason: Empty Profile section\n";
         writeLog(0, "Load profile failed! Reason: Empty Profile section", LOG_LEVEL_ERROR);
@@ -1218,13 +1237,13 @@ std::string getProfile(RESPONSE_CALLBACK_ARGS)
                 writeLog(0, "Ignoring non-exist profile '" + name + "'...", LOG_LEVEL_WARNING);
                 continue;
             }
-            if(ini.ParseFile(name) != INIREADER_EXCEPTION_NONE && !ini.SectionExist("Profile"))
+            if(ini.parse_file(name) != INIREADER_EXCEPTION_NONE && !ini.section_exist("Profile"))
             {
                 writeLog(0, "Ignoring broken profile '" + name + "'...", LOG_LEVEL_WARNING);
                 continue;
             }
-            url = ini.Get("Profile", "url");
-            if(url.size())
+            url = ini.get("Profile", "url");
+            if(!url.empty())
             {
                 all_urls += "|" + url;
                 writeLog(0, "Profile url from '" + name + "' added.", LOG_LEVEL_INFO);
@@ -1238,10 +1257,9 @@ std::string getProfile(RESPONSE_CALLBACK_ARGS)
     }
 
     contents.emplace("token", token);
-    contents.emplace("profile_data", base64Encode(global.managedConfigPrefix + "/getprofile?" + argument));
-    std::string query = std::accumulate(contents.begin(), contents.end(), std::string(), [](const std::string &x, auto y){ return x + y.first + "=" + urlEncode(y.second) + "&"; });
-    query += argument;
-    request.argument = query;
+    contents.emplace("profile_data", base64Encode(global.managedConfigPrefix + "/getprofile?" + joinArguments(argument)));
+    std::copy(argument.cbegin(), argument.cend(), std::inserter(contents, contents.end()));
+    request.argument = contents;
     return subconverter(request, response);
 }
 
@@ -1266,7 +1284,7 @@ inline std::string intToStream(unsigned long long stream)
         streamval /= 1024.0;
     }
     snprintf(chrs, 15, "%.2f %cB", streamval, units[level]);
-    return std::string(chrs);
+    return {chrs};
 }
 
 std::string subInfoToMessage(std::string subinfo)
@@ -1276,7 +1294,7 @@ std::string subInfoToMessage(std::string subinfo)
     std::string retdata, useddata = "N/A", totaldata = "N/A", expirydata = "N/A";
     std::string upload = getUrlArg(subinfo, "upload"), download = getUrlArg(subinfo, "download"), total = getUrlArg(subinfo, "total"), expire = getUrlArg(subinfo, "expire");
     ull used = to_number<ull>(upload, 0) + to_number<ull>(download, 0), tot = to_number<ull>(total, 0);
-    time_t expiry = to_number<time_t>(expire, 0);
+    auto expiry = to_number<time_t>(expire, 0);
     if(used != 0)
         useddata = intToStream(used);
     if(tot != 0)
@@ -1299,7 +1317,7 @@ int simpleGenerator()
 {
     //std::cerr<<"\nReading generator configuration...\n";
     writeLog(0, "Reading generator configuration...", LOG_LEVEL_INFO);
-    std::string config = fileGet("generate.ini"), path, profile, arguments, content;
+    std::string config = fileGet("generate.ini"), path, profile, content;
     if(config.empty())
     {
         //std::cerr<<"Generator configuration not found or empty!\n";
@@ -1308,17 +1326,17 @@ int simpleGenerator()
     }
 
     INIReader ini;
-    if(ini.Parse(config) != INIREADER_EXCEPTION_NONE)
+    if(ini.parse(config) != INIREADER_EXCEPTION_NONE)
     {
-        //std::cerr<<"Generator configuration broken! Reason:"<<ini.GetLastError()<<"\n";
-        writeLog(0, "Generator configuration broken! Reason:" + ini.GetLastError(), LOG_LEVEL_ERROR);
+        //std::cerr<<"Generator configuration broken! Reason:"<<ini.get_last_error()<<"\n";
+        writeLog(0, "Generator configuration broken! Reason:" + ini.get_last_error(), LOG_LEVEL_ERROR);
         return -2;
     }
     //std::cerr<<"Read generator configuration completed.\n\n";
     writeLog(0, "Read generator configuration completed.\n", LOG_LEVEL_INFO);
 
-    string_array sections = ini.GetSections();
-    if(global.generateProfiles.size())
+    string_array sections = ini.get_section_names();
+    if(!global.generateProfiles.empty())
     {
         //std::cerr<<"Generating with specific artifacts: \""<<gen_profile<<"\"...\n";
         writeLog(0, "Generating with specific artifacts: \"" + global.generateProfiles + "\"...", LOG_LEVEL_INFO);
@@ -1344,34 +1362,35 @@ int simpleGenerator()
 
     string_multimap allItems;
     std::string proxy = parseProxy(global.proxySubscription);
-    Request request;
-    Response response;
     for(std::string &x : sections)
     {
-        arguments.clear();
+        Request request;
+        Response response;
         response.status_code = 200;
         //std::cerr<<"Generating artifact '"<<x<<"'...\n";
         writeLog(0, "Generating artifact '" + x + "'...", LOG_LEVEL_INFO);
-        ini.EnterSection(x);
-        if(ini.ItemExist("path"))
-            path = ini.Get("path");
+        ini.enter_section(x);
+        if(ini.item_exist("path"))
+            path = ini.get("path");
         else
         {
             //std::cerr<<"Artifact '"<<x<<"' output path missing! Skipping...\n\n";
             writeLog(0, "Artifact '" + x + "' output path missing! Skipping...\n", LOG_LEVEL_ERROR);
             continue;
         }
-        if(ini.ItemExist("profile"))
+        if(ini.item_exist("profile"))
         {
-            profile = ini.Get("profile");
-            request.argument = "name=" + urlEncode(profile) + "&token=" + global.accessToken + "&expand=true";
+            profile = ini.get("profile");
+            request.argument.emplace("name", profile);
+            request.argument.emplace("token", global.accessToken);
+            request.argument.emplace("expand", "true");
             content = getProfile(request, response);
         }
         else
         {
-            if(ini.GetBool("direct") == true)
+            if(ini.get_bool("direct"))
             {
-                std::string url = ini.Get("url");
+                std::string url = ini.get("url");
                 content = fetchFile(url, proxy, global.cacheSubscription);
                 if(content.empty())
                 {
@@ -1384,16 +1403,14 @@ int simpleGenerator()
                 fileWrite(path, "\xEF\xBB\xBF" + content, true);
                 continue;
             }
-            ini.GetItems(allItems);
+            ini.get_items(allItems);
             allItems.emplace("expand", "true");
             for(auto &y : allItems)
             {
                 if(y.first == "path")
                     continue;
-                arguments += y.first + "=" + urlEncode(y.second) + "&";
+                request.argument.emplace(y.first, y.second);
             }
-            arguments.erase(arguments.size() - 1);
-            request.argument = arguments;
             content = subconverter(request, response);
         }
         if(response.status_code != 200)
@@ -1419,10 +1436,10 @@ int simpleGenerator()
 
 std::string renderTemplate(RESPONSE_CALLBACK_ARGS)
 {
-    std::string &argument = request.argument;
+    auto &argument = request.argument;
     int *status_code = &response.status_code;
 
-    std::string path = urlDecode(getUrlArg(argument, "path"));
+    std::string path = getUrlArg(argument, "path");
     writeLog(0, "Trying to render template '" + path + "'...", LOG_LEVEL_INFO);
 
     if(!startsWith(path, global.templatePath) || !fileExist(path))
@@ -1440,16 +1457,10 @@ std::string renderTemplate(RESPONSE_CALLBACK_ARGS)
     tpl_args.global_vars = global.templateVars;
 
     //load request arguments as template variables
-    string_array req_args = split(argument, "&");
-    string_size pos;
     string_map req_arg_map;
-    for(std::string &x : req_args)
+    for (auto &x : argument)
     {
-        pos = x.find("=");
-        if(pos == x.npos)
-            req_arg_map[x] = "";
-        else
-            req_arg_map[x.substr(0, pos)] = x.substr(pos + 1);
+        req_arg_map[x.first] = x.second;
     }
     tpl_args.request_params = req_arg_map;
 
