@@ -5,9 +5,9 @@
 #include <map>
 #include <inja.hpp>
 
-#include "../config/ruleset.h"
-#include "../generator/config/subexport.h"
-#include "../server/webserver.h"
+#include "config/ruleset.h"
+#include "generator/config/subexport.h"
+#include "server/webserver.h"
 
 std::string parseProxy(const std::string &source);
 
@@ -16,7 +16,6 @@ void readConf();
 int simpleGenerator();
 std::string convertRuleset(const std::string &content, int type);
 
-std::string getConvertedRuleset(RESPONSE_CALLBACK_ARGS);
 std::string getProfile(RESPONSE_CALLBACK_ARGS);
 std::string getRuleset(RESPONSE_CALLBACK_ARGS);
 
