@@ -1,7 +1,7 @@
-#include "../handler/settings.h"
+#include "handler/settings.h"
 #include <string>
 
 Settings global;
 
 bool fileExist(const std::string&, bool) { return false; }
-std::string fileGet(const std::string&, bool) { return std::string(); }
+std::string fileGet(const std::string&, bool) { return ""; }
