@@ -2,14 +2,16 @@
 #define PROXY_H_INCLUDED
 
 #include <string>
+#include <vector>
 
-#include "../../utils/tribool.h"
+#include "utils/tribool.h"
 
 using String = std::string;
+using StringArray = std::vector<String>;
 
-enum ProxyType
+enum class ProxyType
 {
-    Unknow,
+    Unknown,
     Shadowsocks,
     ShadowsocksR,
     VMess,
@@ -17,10 +19,13 @@ enum ProxyType
     Snell,
     HTTP,
     HTTPS,
-    SOCKS5
+    SOCKS5,
+    WireGuard,
+    Hysteria,
+    Hysteria2
 };
 
-inline String getProxyTypeName(int type)
+inline String getProxyTypeName(ProxyType type)
 {
     switch(type)
     {
@@ -40,6 +45,12 @@ inline String getProxyTypeName(int type)
         return "HTTPS";
     case ProxyType::SOCKS5:
         return "SOCKS5";
+    case ProxyType::WireGuard:
+        return "WireGuard";
+    case ProxyType::Hysteria:
+        return "Hysteria";
+    case ProxyType::Hysteria2:
+        return "Hysteria2";
     default:
         return "Unknown";
     }
@@ -47,7 +58,7 @@ inline String getProxyTypeName(int type)
 
 struct Proxy
 {
-    int Type = ProxyType::Unknow;
+    ProxyType Type = ProxyType::Unknown;
     uint32_t Id = 0;
     uint32_t GroupId = 0;
     String Group;
@@ -82,8 +93,40 @@ struct Proxy
     tribool AllowInsecure;
     tribool TLS13;
 
+    String UnderlyingProxy;
+
     uint16_t SnellVersion = 0;
     String ServerName;
+
+    String SelfIP;
+    String SelfIPv6;
+    String PublicKey;
+    String PrivateKey;
+    String PreSharedKey;
+    StringArray DnsServers;
+    uint16_t Mtu = 0;
+    String AllowedIPs = "0.0.0.0/0, ::/0";
+    uint16_t KeepAlive = 0;
+    String TestUrl;
+    String ClientId;
+
+    String Ports;
+    String Up;
+    uint32_t UpSpeed;
+    String Down;
+    uint32_t DownSpeed;
+    String AuthStr;
+    String SNI;
+    String Fingerprint;
+    String Ca;
+    String CaStr;
+    uint32_t RecvWindowConn;
+    uint32_t RecvWindow;
+    tribool DisableMtuDiscovery;
+    uint32_t HopInterval;
+    StringArray Alpn;
+
+    uint32_t CWND = 0;
 };
 
 #define SS_DEFAULT_GROUP "SSProvider"
@@ -93,5 +136,8 @@ struct Proxy
 #define HTTP_DEFAULT_GROUP "HTTPProvider"
 #define TROJAN_DEFAULT_GROUP "TrojanProvider"
 #define SNELL_DEFAULT_GROUP "SnellProvider"
+#define WG_DEFAULT_GROUP "WireGuardProvider"
+#define HYSTERIA_DEFAULT_GROUP "HysteriaProvider"
+#define HYSTERIA2_DEFAULT_GROUP "Hysteria2Provider"
 
 #endif // PROXY_H_INCLUDED
