@@ -3,9 +3,9 @@
 
 #include <string>
 
-#include "../utils/string.h"
+#include "utils/string.h"
 #include "config/proxy.h"
-#include "../config/regmatch.h"
+#include "config/regmatch.h"
 
 bool getSubInfoFromHeader(const std::string &header, std::string &result);
 bool getSubInfoFromNodes(const std::vector<Proxy> &nodes, const RegexMatchConfigs &stream_rules, const RegexMatchConfigs &time_rules, std::string &result);
