@@ -4,18 +4,17 @@
 #include <string>
 #include <map>
 #include <atomic>
-#include <event2/http.h>
 #include <curl/curlver.h>
 
-#include "../utils/map_extra.h"
-#include "../utils/string.h"
-#include "../version.h"
+#include "utils/map_extra.h"
+#include "utils/string.h"
+#include "version.h"
 
 struct Request
 {
     std::string method;
     std::string url;
-    std::string argument;
+    string_multimap argument;
     string_icase_map headers;
     std::string postdata;
 };
@@ -46,7 +45,7 @@ struct responseRoute
     std::string method;
     std::string path;
     std::string content_type;
-    response_callback rc;
+    response_callback rc {};
 };
 
 class WebServer
@@ -63,16 +62,31 @@ public:
     bool require_auth = false;
     std::string auth_user, auth_password, auth_realm = "Please enter username and password:";
 
-    void append_response(const std::string &method, const std::string &uri, const std::string &content_type, response_callback response);
-    void append_redirect(const std::string &uri, const std::string &target);
-    void reset_redirect();
-    int start_web_server(void *argv);
-    int start_web_server_multi(void *argv);
     void stop_web_server();
-private:
-    int serveFile(const std::string &filename, std::string &content_type, std::string &return_data);
-    inline int process_request(Request &request, Response &response, std::string &return_data);
-    void on_request(evhttp_request *req, void *args);
+
+    void append_response(const std::string &method, const std::string &uri, const std::string &content_type, response_callback response)
+    {
+        responseRoute rr;
+        rr.method = method;
+        rr.path = uri;
+        rr.content_type = content_type;
+        rr.rc = response;
+        responses.emplace_back(std::move(rr));
+    }
+
+    void append_redirect(const std::string &uri, const std::string &target)
+    {
+        redirect_map[uri] = target;
+    }
+
+    void reset_redirect()
+    {
+        std::map<std::string, std::string>().swap(redirect_map);
+    }
+
+    int start_web_server(listener_args *args);
+    int start_web_server_multi(listener_args *args);
+
     std::vector<responseRoute> responses;
     string_map redirect_map;
 };
