@@ -7,9 +7,9 @@
 #include <numeric>
 #include <algorithm>
 
-#include "../codepage.h"
-#include "../file_extra.h"
-#include "../string.h"
+#include "utils/codepage.h"
+#include "utils/file_extra.h"
+#include "utils/string.h"
 
 enum
 {
@@ -21,13 +21,12 @@ enum
     INIREADER_EXCEPTION_NONE
 };
 
-using ini_data_struct = std::map<std::string, std::multimap<std::string, std::string>>;
-using string_multimap = std::multimap<std::string, std::string>;
-using string_array = std::vector<std::string>;
-using string_size = std::string::size_type;
-
 class INIReader
 {
+    using ini_data_struct = std::map<std::string, std::multimap<std::string, std::string>>;
+    using string_multimap = std::multimap<std::string, std::string>;
+    using string_array = std::vector<std::string>;
+    using string_size = std::string::size_type;
     /**
     *  @brief A simple INI reader which utilize map and vector
     *  to store sections and items, allowing access in logarithmic time.
@@ -51,17 +50,17 @@ private:
     int last_error = INIREADER_EXCEPTION_NONE;
     unsigned int last_error_index = 0;
 
-    inline int __priv_save_error_and_return(int x)
+    inline int save_error_and_return(int x)
     {
         last_error = x;
         return last_error;
     }
 
-    inline bool __priv_chk_ignore(const std::string &section)
+    inline bool chk_ignore(const std::string &section)
     {
         bool excluded = false, included = false;
         excluded = std::find(exclude_sections.begin(), exclude_sections.end(), section) != exclude_sections.end();
-        if(include_sections.size())
+        if(!include_sections.empty())
             included = std::find(include_sections.begin(), include_sections.end(), section) != include_sections.end();
         else
             included = true;
@@ -69,12 +68,12 @@ private:
         return excluded || !included;
     }
 
-    inline bool __priv_chk_direct_save(const std::string &section)
+    inline bool chk_direct_save(const std::string &section)
     {
         return std::find(direct_save_sections.cbegin(), direct_save_sections.cend(), section) != direct_save_sections.cend();
     }
 
-    inline std::string __priv_get_err_str(int error)
+    inline std::string get_err_str(int error)
     {
         switch(error)
         {
@@ -87,30 +86,30 @@ private:
         case INIREADER_EXCEPTION_OUTOFBOUND:
             return "Item exists outside of any section";
         case INIREADER_EXCEPTION_NOTPARSED:
-            return "Parse error";
+            return "parse error";
         default:
             return "Undefined";
         }
     }
 
-    template <typename T> inline void eraseElements(std::vector<T> &target)
+    template <typename T> inline void erase_elements(std::vector<T> &target)
     {
         target.clear();
         target.shrink_to_fit();
     }
 
-    template <typename T> inline void eraseElements(T &target)
+    template <typename T> inline void erase_elements(T &target)
     {
         T().swap(target);
     }
 public:
     /**
-    *  @brief Set this flag to true to do a UTF8-To-GBK conversion before parsing data. Only useful in Windows.
+    *  @brief set this flag to true to do a UTF8-To-GBK conversion before parsing data. Only useful in Windows.
     */
     bool do_utf8_to_gbk = false;
 
     /**
-    *  @brief Set this flag to true so any line within the section will be stored even it doesn't follow the "name=value" format.
+    *  @brief set this flag to true so any line within the section will be stored even it doesn't follow the "name=value" format.
     * These lines will store as the name "{NONAME}".
     */
     bool store_any_line = false;
@@ -139,12 +138,12 @@ public:
     }
 
     /**
-    *  @brief Parse a file during initialization.
+    *  @brief parse a file during initialization.
     */
     explicit INIReader(const std::string &filePath)
     {
         parsed = false;
-        ParseFile(filePath);
+        parse_file(filePath);
     }
 
     ~INIReader() = default;
@@ -171,18 +170,18 @@ public:
 
     INIReader(const INIReader &src) = default;
 
-    std::string GetLastError()
+    std::string get_last_error()
     {
         if(parsed)
-            return __priv_get_err_str(last_error);
+            return get_err_str(last_error);
         else
-            return "line " + std::to_string(last_error_index) + ": " + __priv_get_err_str(last_error);
+            return "line " + std::to_string(last_error_index) + ": " + get_err_str(last_error);
     }
 
     /**
     *  @brief Exclude a section with the given name.
     */
-    void ExcludeSection(const std::string &section)
+    void exclude_section(const std::string &section)
     {
         exclude_sections.emplace_back(section);
     }
@@ -190,7 +189,7 @@ public:
     /**
     *  @brief Include a section with the given name.
     */
-    void IncludeSection(const std::string &section)
+    void include_section(const std::string &section)
     {
         include_sections.emplace_back(section);
     }
@@ -198,28 +197,28 @@ public:
     /**
     *  @brief Add a section to the direct-save sections list.
     */
-    void AddDirectSaveSection(const std::string &section)
+    void add_direct_save_section(const std::string &section)
     {
         direct_save_sections.emplace_back(section);
     }
 
     /**
-    *  @brief Set isolated items to given section.
+    *  @brief set isolated items to given section.
     */
-    void SetIsolatedItemsSection(const std::string &section)
+    void set_isolated_items_section(const std::string &section)
     {
         isolated_items_section = section;
     }
 
     /**
-    *  @brief Parse INI content into mapped data structure.
+    *  @brief parse INI content into mapped data structure.
     * If exclude sections are set, these sections will not be stored.
     * If include sections are set, only these sections will be stored.
     */
-    int Parse(std::string content) //parse content into mapped data
+    int parse(std::string content) //parse content into mapped data
     {
-        if(!content.size()) //empty content
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_EMPTY);
+        if(content.empty()) //empty content
+            return save_error_and_return(INIREADER_EXCEPTION_EMPTY);
 
         //remove UTF-8 BOM
         if(content.compare(0, 3, "\xEF\xBB\xBF") == 0)
@@ -232,16 +231,16 @@ public:
         std::stringstream strStrm;
         char delimiter = getLineBreak(content);
 
-        EraseAll(); //first erase all data
+        erase_all(); //first erase all data
         if(do_utf8_to_gbk && isStrUTF8(content))
             content = utf8ToACP(content); //do conversion if flag is set
 
-        if(store_isolated_line && isolated_items_section.size())
+        if(store_isolated_line && !isolated_items_section.empty())
         {
             curSection = isolated_items_section; //items before any section define will be store in this section
             //check this section first
-            inExcludedSection = __priv_chk_ignore(curSection); //check if this section is excluded
-            inDirectSaveSection = __priv_chk_direct_save(curSection); //check if this section requires direct-save
+            inExcludedSection = chk_ignore(curSection); //check if this section is excluded
+            inDirectSaveSection = chk_direct_save(curSection); //check if this section requires direct-save
             inIsolatedSection = true;
         }
         strStrm<<content;
@@ -250,31 +249,31 @@ public:
         {
             last_error_index++;
             strLine = trimWhitespace(strLine);
-            string_size lineSize = strLine.size(), pos_equal = strLine.find("=");
+            string_size lineSize = strLine.size(), pos_equal = strLine.find('=');
             if((!lineSize || strLine[0] == ';' || strLine[0] == '#' || (lineSize >= 2 && strLine[0] == '/' && strLine[1] == '/')) && !inDirectSaveSection) //empty lines and comments are ignored
                 continue;
             processEscapeChar(strLine);
             if(strLine[0] == '[' && strLine[lineSize - 1] == ']') //is a section title
             {
                 thisSection = strLine.substr(1, lineSize - 2); //save section title
-                inExcludedSection = __priv_chk_ignore(thisSection); //check if this section is excluded
-                inDirectSaveSection = __priv_chk_direct_save(thisSection); //check if this section requires direct-save
+                inExcludedSection = chk_ignore(thisSection); //check if this section is excluded
+                inDirectSaveSection = chk_direct_save(thisSection); //check if this section requires direct-save
 
-                if(curSection.size() && (keep_empty_section || itemGroup.size())) //just finished reading a section
+                if(!curSection.empty() && (keep_empty_section || !itemGroup.empty())) //just finished reading a section
                 {
                     if(ini_content.find(curSection) != ini_content.end()) //a section with the same name has been inserted
                     {
-                        if(allow_dup_section_titles || !ini_content.at(curSection).size())
+                        if(allow_dup_section_titles || ini_content.at(curSection).empty())
                         {
                             auto iter = ini_content.at(curSection); //get the existing section
                             iter.merge(itemGroup); //move new items to this section
                         }
-                        else if(ini_content.at(curSection).size())
-                            return __priv_save_error_and_return(INIREADER_EXCEPTION_DUPLICATE); //not allowed, stop
+                        else if(!ini_content.at(curSection).empty())
+                            return save_error_and_return(INIREADER_EXCEPTION_DUPLICATE); //not allowed, stop
                     }
                     else if(!inIsolatedSection || isolated_items_section != thisSection)
                     {
-                        if(itemGroup.size())
+                        if(!itemGroup.empty())
                             read_sections.push_back(curSection); //add to read sections list
                         if(std::find(section_order.cbegin(), section_order.cend(), curSection) == section_order.cend())
                             section_order.emplace_back(curSection); //add to section order if not added before
@@ -282,33 +281,33 @@ public:
                     }
                 }
                 inIsolatedSection = false;
-                eraseElements(itemGroup); //reset section storage
+                erase_elements(itemGroup); //reset section storage
                 curSection = thisSection; //start a new section
             }
-            else if(((store_any_line && pos_equal == strLine.npos) || inDirectSaveSection) && !inExcludedSection && curSection.size()) //store a line without name
+            else if(((store_any_line && pos_equal == std::string::npos) || inDirectSaveSection) && !inExcludedSection && !curSection.empty()) //store a line without name
             {
                 itemGroup.emplace("{NONAME}", strLine);
             }
-            else if(pos_equal != strLine.npos) //is an item
+            else if(pos_equal != std::string::npos) //is an item
             {
                 if(inExcludedSection) //this section is excluded
                     continue;
-                if(!curSection.size()) //not in any section
-                    return __priv_save_error_and_return(INIREADER_EXCEPTION_OUTOFBOUND);
+                if(curSection.empty()) //not in any section
+                    return save_error_and_return(INIREADER_EXCEPTION_OUTOFBOUND);
                 string_size pos_value = strLine.find_first_not_of(' ', pos_equal + 1);
                 itemName = trim(strLine.substr(0, pos_equal));
-                if(pos_value != strLine.npos) //not a key with empty value
+                if(pos_value != std::string::npos) //not a key with empty value
                 {
                     itemVal = strLine.substr(pos_value);
                     itemGroup.emplace(std::move(itemName), std::move(itemVal)); //insert to current section
                 }
                 else
-                    itemGroup.emplace(std::move(itemName), std::string());
+                    itemGroup.emplace(std::move(itemName), "");
             }
-            if(include_sections.size() && include_sections == read_sections) //all included sections has been read
+            if(!include_sections.empty() && include_sections == read_sections) //all included sections has been read
                 break; //exit now
         }
-        if(curSection.size() && (keep_empty_section || itemGroup.size())) //final section
+        if(!curSection.empty() && (keep_empty_section || !itemGroup.empty())) //final section
         {
             if(ini_content.find(curSection) != ini_content.end()) //a section with the same name has been inserted
             {
@@ -317,12 +316,12 @@ public:
                     auto &iter = ini_content.at(curSection); //get the existing section
                     iter.merge(itemGroup); //move new items to this section
                 }
-                else if(ini_content.at(curSection).size())
-                    return __priv_save_error_and_return(INIREADER_EXCEPTION_DUPLICATE); //not allowed, stop
+                else if(!ini_content.at(curSection).empty())
+                    return save_error_and_return(INIREADER_EXCEPTION_DUPLICATE); //not allowed, stop
             }
             else if(!inIsolatedSection || isolated_items_section != thisSection)
             {
-                if(itemGroup.size())
+                if(!itemGroup.empty())
                     read_sections.emplace_back(curSection); //add to read sections list
                 if(std::find(section_order.cbegin(), section_order.cend(), curSection) == section_order.cend())
                     section_order.emplace_back(curSection); //add to section order if not added before
@@ -330,23 +329,23 @@ public:
             }
         }
         parsed = true;
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE); //all done
+        return save_error_and_return(INIREADER_EXCEPTION_NONE); //all done
     }
 
     /**
-    *  @brief Parse an INI file into mapped data structure.
+    *  @brief parse an INI file into mapped data structure.
     */
-    int ParseFile(const std::string &filePath)
+    int parse_file(const std::string &filePath)
     {
         if(!fileExist(filePath))
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
-        return Parse(fileGet(filePath));
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        return parse(fileGet(filePath));
     }
 
     /**
     *  @brief Check whether a section exist.
     */
-    bool SectionExist(const std::string &section)
+    bool section_exist(const std::string &section)
     {
         return ini_content.find(section) != ini_content.end();
     }
@@ -354,7 +353,7 @@ public:
     /**
     *  @brief Count of sections in the whole INI.
     */
-    unsigned int SectionCount()
+    unsigned int section_count()
     {
         return ini_content.size();
     }
@@ -362,7 +361,7 @@ public:
     /**
     *  @brief Return all section names inside INI.
     */
-    string_array GetSections()
+    string_array get_section_names()
     {
         return section_order;
     }
@@ -370,19 +369,19 @@ public:
     /**
     *  @brief Enter a section with the given name. Section name and data will be cached to speed up the following reading process.
     */
-    int EnterSection(const std::string &section)
+    int enter_section(const std::string &section)
     {
-        if(!SectionExist(section))
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        if(!section_exist(section))
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
         current_section = cached_section = section;
         cached_section_content = ini_content.find(section);
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+        return save_error_and_return(INIREADER_EXCEPTION_NONE);
     }
 
     /**
-    *  @brief Set current section.
+    *  @brief set current section.
     */
-    void SetCurrentSection(const std::string &section)
+    void set_current_section(const std::string &section)
     {
         current_section = section;
     }
@@ -390,9 +389,9 @@ public:
     /**
     *  @brief Check whether an item exist in the given section. Return false if the section does not exist.
     */
-    bool ItemExist(const std::string &section, const std::string &itemName)
+    bool item_exist(const std::string &section, const std::string &itemName)
     {
-        if(!SectionExist(section))
+        if(!section_exist(section))
             return false;
 
         if(section != cached_section)
@@ -407,17 +406,17 @@ public:
     /**
     *  @brief Check whether an item exist in current section. Return false if the section does not exist.
     */
-    bool ItemExist(const std::string &itemName)
+    bool item_exist(const std::string &itemName)
     {
-        return current_section.size() ? ItemExist(current_section, itemName) : false;
+        return !current_section.empty() && item_exist(current_section, itemName);
     }
 
     /**
     *  @brief Check whether an item with the given name prefix exist in the given section. Return false if the section does not exist.
     */
-    bool ItemPrefixExist(const std::string &section, const std::string &itemName)
+    bool item_prefix_exists(const std::string &section, const std::string &itemName)
     {
-        if(!SectionExist(section))
+        if(!section_exist(section))
             return false;
 
         if(section != cached_section)
@@ -426,49 +425,47 @@ public:
             cached_section_content = ini_content.find(section);
         }
 
-        for(auto &x : cached_section_content->second)
-        {
-            if(x.first.find(itemName) == 0)
-                return true;
-        }
+        auto &items = cached_section_content->second;
 
-        return false;
+        return std::any_of(items.cbegin(), items.cend(), [&](auto &x) {
+            return x.first.find(itemName) == 0;
+        });
     }
 
     /**
     *  @brief Check whether an item with the given name prefix exist in current section. Return false if the section does not exist.
     */
-    bool ItemPrefixExist(const std::string &itemName)
+    bool item_prefix_exist(const std::string &itemName)
     {
-        return current_section.size() ? ItemPrefixExist(current_section, itemName) : false;
+        return !current_section.empty() && item_prefix_exists(current_section, itemName);
     }
 
     /**
     *  @brief Count of items in the given section. Return 0 if the section does not exist.
     */
-    unsigned int ItemCount(const std::string &section)
+    unsigned int item_count(const std::string &section)
     {
-        if(!parsed || !SectionExist(section))
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
+        if(!parsed || !section_exist(section))
+            return save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
 
         return ini_content.at(section).size();
     }
 
     /**
-    *  @brief Erase all data from the data structure and reset parser status.
+    *  @brief erase all data from the data structure and reset parser status.
     */
-    void EraseAll()
+    void erase_all()
     {
-        eraseElements(ini_content);
-        eraseElements(section_order);
+        erase_elements(ini_content);
+        erase_elements(section_order);
         cached_section.clear();
         cached_section_content = ini_content.end();
         parsed = false;
     }
 
-    ini_data_struct::iterator GetItemsRef(const std::string &section)
+    ini_data_struct::iterator get_items_ref(const std::string &section)
     {
-        if(!parsed || !SectionExist(section))
+        if(!parsed || !section_exist(section))
             return ini_content.end();
 
         if(cached_section != section)
@@ -482,35 +479,35 @@ public:
     /**
     *  @brief Retrieve all items in the given section.
     */
-    int GetItems(const std::string &section, string_multimap &data)
+    int get_items(const std::string &section, string_multimap &data)
     {
-        auto section_ref = GetItemsRef(section);
+        auto section_ref = get_items_ref(section);
         if(section_ref == ini_content.end())
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
 
         data = section_ref->second;
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+        return save_error_and_return(INIREADER_EXCEPTION_NONE);
     }
 
     /**
     *  @brief Retrieve all items in current section.
     */
-    int GetItems(string_multimap &data)
+    int get_items(string_multimap &data)
     {
-        return current_section.size() ? GetItems(current_section, data) : -1;
+        return !current_section.empty() ? get_items(current_section, data) : -1;
     }
 
     /**
     * @brief Retrieve item(s) with the same name prefix in the given section.
     */
-    int GetAll(const std::string &section, const std::string &itemName, string_array &results) //retrieve item(s) with the same itemName prefix
+    int get_all(const std::string &section, const std::string &itemName, string_array &results) //retrieve item(s) with the same itemName prefix
     {
         if(!parsed)
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
+            return save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
 
-        auto section_ref = GetItemsRef(section);
+        auto section_ref = get_items_ref(section);
         if(section_ref == ini_content.end())
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
 
         for(auto &x : section_ref->second)
         {
@@ -518,24 +515,24 @@ public:
                 results.emplace_back(x.second);
         }
 
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+        return save_error_and_return(INIREADER_EXCEPTION_NONE);
     }
 
     /**
     * @brief Retrieve item(s) with the same name prefix in current section.
     */
-    int GetAll(const std::string &itemName, string_array &results)
+    int get_all(const std::string &itemName, string_array &results)
     {
-        return current_section.size() ? GetAll(current_section, itemName, results) : -1;
+        return !current_section.empty() ? get_all(current_section, itemName, results) : -1;
     }
 
     /**
     * @brief Retrieve one item with the exact same name in the given section.
     */
-    std::string Get(const std::string &section, const std::string &itemName) //retrieve one item with the exact same itemName
+    std::string get(const std::string &section, const std::string &itemName) //retrieve one item with the exact same itemName
     {
-        if(!parsed || !SectionExist(section))
-            return std::string();
+        if(!parsed || !section_exist(section))
+            return "";
 
         if(cached_section != section)
         {
@@ -548,68 +545,68 @@ public:
         if(iter != cache.end())
             return iter->second;
 
-        return std::string();
+        return "";
     }
 
     /**
     * @brief Retrieve one item with the exact same name in current section.
     */
-    std::string Get(const std::string &itemName)
+    std::string get(const std::string &itemName)
     {
-        return current_section.size() ? Get(current_section, itemName) : std::string();
+        return !current_section.empty() ? get(current_section, itemName) : "";
     }
 
     /**
     * @brief Retrieve one item with the exact same name in the given section, if exist.
     */
-    int GetIfExist(const std::string &section, const std::string &itemName, std::string &target) //retrieve one item with the exact same itemName
+    int get_if_exist(const std::string &section, const std::string &itemName, std::string &target) //retrieve one item with the exact same itemName
     {
         if(!parsed)
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
+            return save_error_and_return(INIREADER_EXCEPTION_NOTPARSED);
 
-        if(ItemExist(section, itemName))
+        if(item_exist(section, itemName))
         {
-            target = Get(section, itemName);
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+            target = get(section, itemName);
+            return save_error_and_return(INIREADER_EXCEPTION_NONE);
         }
 
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
     }
 
     /**
     * @brief Retrieve one item with the exact same name in current section, if exist.
     */
-    int GetIfExist(const std::string &itemName, std::string &target)
+    int get_if_exist(const std::string &itemName, std::string &target)
     {
-        return current_section.size() ? GetIfExist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
+        return !current_section.empty() ? get_if_exist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
     }
 
     /**
     * @brief Retrieve one boolean item value with the exact same name in the given section.
     */
-    bool GetBool(const std::string &section, const std::string &itemName)
+    bool get_bool(const std::string &section, const std::string &itemName)
     {
-        return Get(section, itemName) == "true";
+        return get(section, itemName) == "true";
     }
 
     /**
     * @brief Retrieve one boolean item value with the exact same name in current section.
     */
-    bool GetBool(const std::string &itemName)
+    bool get_bool(const std::string &itemName)
     {
-        return current_section.size() ? Get(current_section, itemName) == "true" : false;
+        return !current_section.empty() && get(current_section, itemName) == "true";
     }
 
     /**
     * @brief Retrieve one boolean item value with the exact same name in the given section.
     */
-    int GetBoolIfExist(const std::string &section, const std::string &itemName, bool &target)
+    int get_bool_if_exist(const std::string &section, const std::string &itemName, bool &target)
     {
         std::string result;
-        int retval = GetIfExist(section, itemName, result);
-        if(retval != INIREADER_EXCEPTION_NONE)
-            return retval;
-        if(result.size())
+        int retVal = get_if_exist(section, itemName, result);
+        if(retVal != INIREADER_EXCEPTION_NONE)
+            return retVal;
+        if(!result.empty())
             target = result == "true";
         return INIREADER_EXCEPTION_NONE;
     }
@@ -617,21 +614,21 @@ public:
     /**
     * @brief Retrieve one boolean item value with the exact same name in current section.
     */
-    int GetBoolIfExist(const std::string &itemName, bool &target)
+    int get_bool_if_exist(const std::string &itemName, bool &target)
     {
-        return current_section.size() ? GetBoolIfExist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
+        return !current_section.empty() ? get_bool_if_exist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
     }
 
     /**
     * @brief Retrieve one number item value with the exact same name in the given section.
     */
-    template <typename T> int GetNumberIfExist(const std::string &section, const std::string &itemName, T &target)
+    template <typename T> int get_number_if_exist(const std::string &section, const std::string &itemName, T &target)
     {
         std::string result;
-        int retval = GetIfExist(section, itemName, result);
-        if(retval != INIREADER_EXCEPTION_NONE)
-            return retval;
-        if(result.size())
+        int retVal = get_if_exist(section, itemName, result);
+        if(retVal != INIREADER_EXCEPTION_NONE)
+            return retVal;
+        if(!result.empty())
             target = to_number<T>(result, target);
         return INIREADER_EXCEPTION_NONE;
     }
@@ -639,73 +636,72 @@ public:
     /**
     * @brief Retrieve one number item value with the exact same name in current section.
     */
-    template <typename T> int GetNumberIfExist(const std::string &itemName, T &target)
+    template <typename T> int get_number_if_exist(const std::string &itemName, T &target)
     {
-        return current_section.size() ? GetNumberIfExist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
+        return !current_section.empty() ? get_number_if_exist(current_section, itemName, target) : INIREADER_EXCEPTION_NOTEXIST;
     }
 
     /**
     * @brief Retrieve one integer item value with the exact same name in the given section.
     */
-    int GetIntIfExist(const std::string &section, const std::string &itemName, int &target)
+    int get_int_if_exist(const std::string &section, const std::string &itemName, int &target)
     {
-        return GetNumberIfExist<int>(section, itemName, target);
+        return get_number_if_exist<int>(section, itemName, target);
     }
 
     /**
     * @brief Retrieve one integer item value with the exact same name in current section.
     */
-    int GetIntIfExist(const std::string &itemName, int &target)
+    int get_int_if_exist(const std::string &itemName, int &target)
     {
-        return GetNumberIfExist<int>(itemName, target);
+        return get_number_if_exist<int>(itemName, target);
     }
 
     /**
     * @brief Retrieve one integer item value with the exact same name in the given section.
     */
-    int GetInt(const std::string &section, const std::string &itemName)
+    int get_int(const std::string &section, const std::string &itemName)
     {
-        return to_int(Get(section, itemName), 0);
+        return to_int(get(section, itemName), 0);
     }
 
     /**
     * @brief Retrieve one integer item value with the exact same name in current section.
     */
-    int GetInt(const std::string &itemName)
+    int get_int(const std::string &itemName)
     {
-        return GetInt(current_section, itemName);
+        return get_int(current_section, itemName);
     }
 
     /**
     * @brief Retrieve the first item found in the given section.
     */
-    std::string GetFirst(const std::string &section, const std::string &itemName) //return the first item value found in section
+    std::string get_first(const std::string &section, const std::string &itemName) //return the first item value found in section
     {
         if(!parsed)
-            return std::string();
+            return "";
         string_array result;
-        if(GetAll(section, itemName, result) != -1)
+        if(get_all(section, itemName, result) != -1)
             return result[0];
         else
-            return std::string();
+            return "";
     }
 
     /**
     * @brief Retrieve the first item found in current section.
     */
-    std::string GetFirst(const std::string &itemName)
+    std::string get_first(const std::string &itemName)
     {
-        return current_section.size() ? GetFirst(current_section, itemName) : std::string();
+        return !current_section.empty() ? get_first(current_section, itemName) : "";
     }
 
     /**
     * @brief Retrieve a string style array with specific separator and write into integer array.
     */
-    template <typename T> void GetIntArray(const std::string &section, const std::string &itemName, const std::string &separator, T &Array)
+    template <typename T, size_t N> void get_int_array(const std::string &section, const std::string &itemName, const std::string &separator, T& Array)
     {
-        string_array vArray;
-        unsigned int index, UBound = sizeof(Array) / sizeof(Array[0]);
-        vArray = split(Get(section, itemName), separator);
+        unsigned int index, UBound = sizeof(Array[0]) / sizeof(Array);
+        string_array vArray = split(get(section, itemName), separator);
         for(index = 0; index < vArray.size() && index < UBound; index++)
             Array[index] = stoi(vArray[index]);
         for(; index < UBound; index++)
@@ -715,24 +711,24 @@ public:
     /**
     * @brief Retrieve a string style array with specific separator and write into integer array.
     */
-    template <typename T> void GetIntArray(const std::string &itemName, const std::string &separator, T &Array)
+    template <typename T> void get_int_array(const std::string &itemName, const std::string &separator, T& Array)
     {
-        if(current_section.size())
-            GetIntArray(current_section, itemName, separator, Array);
+        if(!current_section.empty())
+            get_int_array(current_section, itemName, separator, Array);
     }
 
     /**
     *  @brief Add a std::string value with given values.
     */
-    int Set(const std::string &section, std::string itemName, std::string itemVal)
+    int set(const std::string &section, std::string itemName, std::string itemVal)
     {
-        if(!section.size())
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        if(section.empty())
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
 
         if(!parsed)
             parsed = true;
 
-        if(SectionExist(section))
+        if(section_exist(section))
         {
             string_multimap &mapTemp = ini_content.at(section);
             mapTemp.insert(std::pair<std::string, std::string>(std::move(itemName), std::move(itemVal)));
@@ -745,110 +741,109 @@ public:
             section_order.emplace_back(section);
         }
 
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+        return save_error_and_return(INIREADER_EXCEPTION_NONE);
     }
 
     /**
     *  @brief Add a string value with given values.
     */
-    int Set(std::string itemName, std::string itemVal)
+    int set(std::string itemName, std::string itemVal)
     {
-        if(!current_section.size())
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
-        return Set(current_section, std::move(itemName), std::move(itemVal));
+        if(current_section.empty())
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        return set(current_section, std::move(itemName), std::move(itemVal));
     }
 
     /**
     *  @brief Add a boolean value with given values.
     */
-    int SetBool(const std::string &section, std::string itemName, bool itemVal)
+    int set_bool(const std::string &section, std::string itemName, bool itemVal)
     {
-        return Set(section, std::move(itemName), itemVal ? "true" : "false");
+        return set(section, std::move(itemName), itemVal ? "true" : "false");
     }
 
     /**
     *  @brief Add a boolean value with given values.
     */
-    int SetBool(std::string itemName, bool itemVal)
+    int set_bool(std::string itemName, bool itemVal)
     {
-        return SetBool(current_section, std::move(itemName), itemVal);
+        return set_bool(current_section, std::move(itemName), itemVal);
     }
 
     /**
     *  @brief Add a double value with given values.
     */
-    int SetDouble(const std::string &section, std::string itemName, double itemVal)
+    int set_double(const std::string &section, std::string itemName, double itemVal)
     {
-        return Set(section, std::move(itemName), std::to_string(itemVal));
+        return set(section, std::move(itemName), std::to_string(itemVal));
     }
 
     /**
     *  @brief Add a double value with given values.
     */
-    int SetDouble(std::string itemName, double itemVal)
+    int set_double(std::string itemName, double itemVal)
     {
-        return SetDouble(current_section, std::move(itemName), itemVal);
+        return set_double(current_section, std::move(itemName), itemVal);
     }
 
     /**
     *  @brief Add a long value with given values.
     */
-    int SetLong(const std::string &section, std::string itemName, long itemVal)
+    int set_long(const std::string &section, std::string itemName, long itemVal)
     {
-        return Set(section, std::move(itemName), std::to_string(itemVal));
+        return set(section, std::move(itemName), std::to_string(itemVal));
     }
 
     /**
     *  @brief Add a long value with given values.
     */
-    int SetLong(std::string itemName, long itemVal)
+    int set_long(std::string itemName, long itemVal)
     {
-        return SetLong(current_section, std::move(itemName), itemVal);
+        return set_long(current_section, std::move(itemName), itemVal);
     }
 
     /**
     *  @brief Add an array with the given separator.
     */
-    template <typename T> int SetArray(const std::string &section, std::string itemName, const std::string &separator, T &Array)
+    template <typename T> int set_array(const std::string &section, std::string itemName, const std::string &separator, T &Array)
     {
         std::string data;
         data = std::accumulate(std::begin(Array), std::end(Array), std::string(), [&](auto a, auto b) { return std::move(a) + std::to_string(b) + separator; });
         data.erase(data.size() - 1);
-        return Set(section, std::move(itemName), data);
+        return set(section, std::move(itemName), data);
     }
 
     /**
     *  @brief Add an array with the given separator.
     */
-    template <typename T> int SetArray(std::string itemName, const std::string &separator, T &Array)
+    template <typename T> int set_array(std::string itemName, const std::string &separator, T &Array)
     {
-        return current_section.size() ? SetArray(current_section, std::move(itemName), separator, Array) : -1;
+        return !current_section.empty() ? set_array(current_section, std::move(itemName), separator, Array) : -1;
     }
 
     /**
     *  @brief Rename an existing section.
     */
-    int RenameSection(const std::string &oldName, std::string newName)
+    int rename_section(const std::string &oldName, const std::string& newName)
     {
-        if(!SectionExist(oldName) || SectionExist(newName))
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_DUPLICATE);
+        if(!section_exist(oldName) || section_exist(newName))
+            return save_error_and_return(INIREADER_EXCEPTION_DUPLICATE);
         auto nodeHandler = ini_content.extract(oldName);
-        nodeHandler.key() = std::move(newName);
+        nodeHandler.key() = newName;
         ini_content.insert(std::move(nodeHandler));
         std::replace(section_order.begin(), section_order.end(), oldName, newName);
-        return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+        return save_error_and_return(INIREADER_EXCEPTION_NONE);
     }
 
     /**
-    *  @brief Erase all items with the given name.
+    *  @brief erase all items with the given name.
     */
-    int Erase(const std::string &section, const std::string &itemName)
+    int erase(const std::string &section, const std::string &itemName)
     {
-        int retVal;
-        if(!SectionExist(section))
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+        if(!section_exist(section))
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
 
-        retVal = ini_content.at(section).erase(itemName);
+        auto retVal = ini_content.at(section).erase(itemName);
         if(retVal && cached_section == section)
         {
             cached_section_content = ini_content.find(section);
@@ -857,47 +852,47 @@ public:
     }
 
     /**
-    *  @brief Erase all items with the given name.
+    *  @brief erase all items with the given name.
     */
-    int Erase(const std::string &itemName)
+    int erase(const std::string &itemName)
     {
-        return current_section.size() ? Erase(current_section, itemName) : -1;
+        return !current_section.empty() ? erase(current_section, itemName) : -1;
     }
 
     /**
-    *  @brief Erase the first item with the given name.
+    *  @brief erase the first item with the given name.
     */
-    int EraseFirst(const std::string &section, const std::string &itemName)
+    int erase_first(const std::string &section, const std::string &itemName)
     {
         string_multimap &mapTemp = ini_content.at(section);
-        string_multimap::iterator iter = mapTemp.find(itemName);
+        auto iter = mapTemp.find(itemName);
         if(iter != mapTemp.end())
         {
             mapTemp.erase(iter);
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NONE);
+            return save_error_and_return(INIREADER_EXCEPTION_NONE);
         }
         else
         {
-            return __priv_save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
+            return save_error_and_return(INIREADER_EXCEPTION_NOTEXIST);
         }
     }
 
     /**
-    *  @brief Erase the first item with the given name.
+    *  @brief erase the first item with the given name.
     */
-    int EraseFirst(const std::string &itemName)
+    int erase_first(const std::string &itemName)
     {
-        return current_section.size() ? EraseFirst(current_section, itemName) : -1;
+        return !current_section.empty() ? erase_first(current_section, itemName) : -1;
     }
 
     /**
-    *  @brief Erase all items in the given section.
+    *  @brief erase all items in the given section.
     */
-    void EraseSection(const std::string &section)
+    void erase_section(const std::string &section)
     {
         if(ini_content.find(section) == ini_content.end())
             return;
-        eraseElements(ini_content.at(section));
+        erase_elements(ini_content.at(section));
         if(cached_section == section)
         {
             cached_section_content = ini_content.end();
@@ -906,18 +901,18 @@ public:
     }
 
     /**
-    *  @brief Erase all items in current section.
+    *  @brief erase all items in current section.
     */
-    void EraseSection()
+    void erase_section()
     {
-        if(current_section.size())
-            EraseSection(current_section);
+        if(!current_section.empty())
+            erase_section(current_section);
     }
 
     /**
     *  @brief Remove a section from INI.
     */
-    void RemoveSection(const std::string &section)
+    void remove_section(const std::string &section)
     {
         if(ini_content.find(section) == ini_content.end())
             return;
@@ -933,21 +928,21 @@ public:
     /**
     *  @brief Remove current section from INI.
     */
-    void RemoveSection()
+    void remove_section()
     {
-        if(current_section.size())
-            RemoveSection(current_section);
+        if(!current_section.empty())
+            remove_section(current_section);
     }
 
     /**
     *  @brief Export the whole INI data structure into a string.
     */
-    std::string ToString()
+    std::string to_string()
     {
         std::string content, itemVal;
 
         if(!parsed)
-            return std::string();
+            return "";
 
         for(auto &x : section_order)
         {
@@ -981,9 +976,9 @@ public:
     /**
     *  @brief Export the whole INI data structure into a file.
     */
-    int ToFile(const std::string &filePath)
+    int to_file(const std::string &filePath)
     {
-        return fileWrite(filePath, ToString(), true);
+        return fileWrite(filePath, to_string(), true);
     }
 };
 
